Euro foreign exchange reference rates

Fetch the latest or historical Euro foreign exchange reference rates from the European Central Bank.

Usage

ecb_euro_rates(x = c("latest", "history"))

Source

http://www.ecb.europa.eu/stats/eurofxref/

Arguments

x

character(1) one of "latest" or "history". Default "latest".

Value

A data.frame() with the reference rates.

Details

Note you can achieve the same by calling the ecb_data() function with the right parameters for each currency.

The reference rates are usually updated at around 16:00 CET every working day, except on TARGET closing days.

They are based on the daily concertation procedure between central banks across Europe, which normally takes place around 14:10 CET. The reference rates are published for information purposes only. Using the rates for transaction purposes is strongly discouraged.

Examples

# \donttest{
ecb_euro_rates()
#> # A tibble: 31 × 3
#>    date       currency    rate
#>    <date>     <chr>      <dbl>
#>  1 2024-09-30 USD        1.12 
#>  2 2024-09-30 JPY      160.   
#>  3 2024-09-30 BGN        1.96 
#>  4 2024-09-30 CZK       25.2  
#>  5 2024-09-30 DKK        7.46 
#>  6 2024-09-30 GBP        0.835
#>  7 2024-09-30 HUF      397.   
#>  8 2024-09-30 PLN        4.28 
#>  9 2024-09-30 RON        4.98 
#> 10 2024-09-30 SEK       11.3  
#> # ℹ 21 more rows
# }