Fetch Deutsche Bundesbank (BBk) data

Fetch Deutsche Bundesbank (BBk) data

Usage

bbk_data(
  flow,
  key = NULL,
  start_period = NULL,
  end_period = NULL,
  first_n = NULL,
  last_n = NULL
)

Source

https://www.bundesbank.de/en/statistics/time-series-databases/help-for-sdmx-web-service/web-service-interface-data

Arguments

flow

(character(1)) flow to query, 5-8 characters. See bbk_metadata() for available dataflows.

key

(character()) key to query.

start_period

(character(1)) start date of the data. Supported formats:

YYYY for annual data (e.g., `2019“)

YYYY-S[1-2] for semi-annual data (e.g., "2019-S1")

YYYY-Q[1-4] for quarterly data (e.g., "2019-Q1")

YYYY-MM for monthly data (e.g., "2019-01")

YYYY-W[01-53] for weekly data (e.g., "2019-W01")

YYYY-MM-DD for daily and business data (e.g., "2019-01-01")

If NULL, no start date restriction is applied (data retrieved from the earliest available date). Default NULL.

end_period

(character(1)) end date of the data, in the same format as start_period. If NULL, no end date restriction is applied (data retrieved up to the most recent available date). Default NULL.

first_n

(numeric(1)) number of observations to retrieve from the start of the series. If NULL, no restriction is applied. Default NULL.

last_n

(numeric(1)) number of observations to retrieve from the end of the series. If NULL, no restriction is applied. Default NULL.

Value

A data.table::data.table() with the requested data.

See also

Examples

# \donttest{
# fetch all data for a given flow and key
bbk_data("BBSIS", "D.I.ZAR.ZI.EUR.S1311.B.A604.R10XX.R.A.A._Z._Z.A", start_period = 2025)
#> Error in req_perform(req_url_query(bbk_build_request(resource), ...)): HTTP 400 Bad Request.
#> If there is a syntactic or semantic issue with the parameters you supplied, a
#> 400 HTTP status code will be returned.
#> See docs at
#> <https://www.bundesbank.de/en/statistics/time-series-databases/help-for-sdmx-web-service/status-codes/status-codes-855918>
# fetch data for multiple keys
bbk_data("BBEX3", c("M.ISK.EUR", "USD.CA.AC.A01"))
#>             date                       key  value
#>           <Date>                    <char>  <num>
#>    1: 1999-01-01 BBEX3.M.ISK.EUR.CA.AC.A01  79.74
#>    2: 1999-02-01 BBEX3.M.ISK.EUR.CA.AC.A01  79.61
#>    3: 1999-03-01 BBEX3.M.ISK.EUR.CA.AC.A01  78.07
#>    4: 1999-04-01 BBEX3.M.ISK.EUR.CA.AC.A01  77.92
#>    5: 1999-05-01 BBEX3.M.ISK.EUR.CA.AC.A01  77.88
#>   ---                                            
#> 1584: 2025-03-01 BBEX3.M.ISK.USD.CA.AC.A01 131.95
#> 1585: 2025-04-01 BBEX3.M.ISK.USD.CA.AC.A01 128.29
#> 1586: 2025-05-01 BBEX3.M.ISK.USD.CA.AC.A01 127.17
#> 1587: 2025-06-01 BBEX3.M.ISK.USD.CA.AC.A01 121.33
#> 1588: 2025-07-01 BBEX3.M.ISK.USD.CA.AC.A01 124.24
#>                                                                        title
#>                                                                       <char>
#>    1:      Exchange rates for the euro in Iceland / EUR 1 = ISK ... (middle)
#>    2:      Exchange rates for the euro in Iceland / EUR 1 = ISK ... (middle)
#>    3:      Exchange rates for the euro in Iceland / EUR 1 = ISK ... (middle)
#>    4:      Exchange rates for the euro in Iceland / EUR 1 = ISK ... (middle)
#>    5:      Exchange rates for the euro in Iceland / EUR 1 = ISK ... (middle)
#>   ---                                                                       
#> 1584: Exchange rates for the US dollar in Iceland / USD 1 = ISK ... (middle)
#> 1585: Exchange rates for the US dollar in Iceland / USD 1 = ISK ... (middle)
#> 1586: Exchange rates for the US dollar in Iceland / USD 1 = ISK ... (middle)
#> 1587: Exchange rates for the US dollar in Iceland / USD 1 = ISK ... (middle)
#> 1588: Exchange rates for the US dollar in Iceland / USD 1 = ISK ... (middle)
#>          freq currency erx_partner_currency erx_series_type erx_rate_type
#>        <char>   <char>               <char>          <char>        <char>
#>    1: monthly      ISK                  EUR              CA            AC
#>    2: monthly      ISK                  EUR              CA            AC
#>    3: monthly      ISK                  EUR              CA            AC
#>    4: monthly      ISK                  EUR              CA            AC
#>    5: monthly      ISK                  EUR              CA            AC
#>   ---                                                                    
#> 1584: monthly      ISK                  USD              CA            AC
#> 1585: monthly      ISK                  USD              CA            AC
#> 1586: monthly      ISK                  USD              CA            AC
#> 1587: monthly      ISK                  USD              CA            AC
#> 1588: monthly      ISK                  USD              CA            AC
#>       erx_suffix time_format decimals   unit unit_mult category
#>           <char>      <char>    <int> <char>    <char>   <char>
#>    1:        A01         P1M        2    ISK         0     WEDE
#>    2:        A01         P1M        2    ISK         0     WEDE
#>    3:        A01         P1M        2    ISK         0     WEDE
#>    4:        A01         P1M        2    ISK         0     WEDE
#>    5:        A01         P1M        2    ISK         0     WEDE
#>   ---                                                          
#> 1584:        A01         P1M        4    ISK         0     WEDE
#> 1585:        A01         P1M        4    ISK         0     WEDE
#> 1586:        A01         P1M        4    ISK         0     WEDE
#> 1587:        A01         P1M        4    ISK         0     WEDE
#> 1588:        A01         P1M        4    ISK         0     WEDE
#>                                                                                                                                         comm_gen_eng
#>                                                                                                                                               <char>
#>    1: Collapse of banking system and introduction of capital controls in October 2008, see Deutsche Bundesbank, Monthly Report, July 2010, pp 52-53.
#>    2: Collapse of banking system and introduction of capital controls in October 2008, see Deutsche Bundesbank, Monthly Report, July 2010, pp 52-53.
#>    3: Collapse of banking system and introduction of capital controls in October 2008, see Deutsche Bundesbank, Monthly Report, July 2010, pp 52-53.
#>    4: Collapse of banking system and introduction of capital controls in October 2008, see Deutsche Bundesbank, Monthly Report, July 2010, pp 52-53.
#>    5: Collapse of banking system and introduction of capital controls in October 2008, see Deutsche Bundesbank, Monthly Report, July 2010, pp 52-53.
#>   ---                                                                                                                                               
#> 1584: Collapse of banking system and introduction of capital controls in October 2008, see Deutsche Bundesbank, Monthly Report, July 2010, pp 52-53.
#> 1585: Collapse of banking system and introduction of capital controls in October 2008, see Deutsche Bundesbank, Monthly Report, July 2010, pp 52-53.
#> 1586: Collapse of banking system and introduction of capital controls in October 2008, see Deutsche Bundesbank, Monthly Report, July 2010, pp 52-53.
#> 1587: Collapse of banking system and introduction of capital controls in October 2008, see Deutsche Bundesbank, Monthly Report, July 2010, pp 52-53.
#> 1588: Collapse of banking system and introduction of capital controls in October 2008, see Deutsche Bundesbank, Monthly Report, July 2010, pp 52-53.
#>                                                                comm_meth_eng
#>                                                                       <char>
#>    1: Until March 2020, calculated on the basis of buying and selling rates.
#>    2: Until March 2020, calculated on the basis of buying and selling rates.
#>    3: Until March 2020, calculated on the basis of buying and selling rates.
#>    4: Until March 2020, calculated on the basis of buying and selling rates.
#>    5: Until March 2020, calculated on the basis of buying and selling rates.
#>   ---                                                                       
#> 1584: Until March 2020, calculated on the basis of buying and selling rates.
#> 1585: Until March 2020, calculated on the basis of buying and selling rates.
#> 1586: Until March 2020, calculated on the basis of buying and selling rates.
#> 1587: Until March 2020, calculated on the basis of buying and selling rates.
#> 1588: Until March 2020, calculated on the basis of buying and selling rates.
#>                         comm_src_eng
#>                               <char>
#>    1: Sedlabanki Islands, Reykjavik.
#>    2: Sedlabanki Islands, Reykjavik.
#>    3: Sedlabanki Islands, Reykjavik.
#>    4: Sedlabanki Islands, Reykjavik.
#>    5: Sedlabanki Islands, Reykjavik.
#>   ---                               
#> 1584: Sedlabanki Islands, Reykjavik.
#> 1585: Sedlabanki Islands, Reykjavik.
#> 1586: Sedlabanki Islands, Reykjavik.
#> 1587: Sedlabanki Islands, Reykjavik.
#> 1588: Sedlabanki Islands, Reykjavik.
# specified period (start date-end date) for daily data
bbk_data(
  "BBSIS", "D.I.ZAR.ZI.EUR.S1311.B.A604.R10XX.R.A.A._Z._Z.A",
  start_period = "2020-01-01",
  end_period = "2020-08-01"
)
#>            date                                                   key value
#>          <Date>                                                <char> <num>
#>   1: 2020-01-02 BBSIS.D.I.ZAR.ZI.EUR.S1311.B.A604.R10XX.R.A.A._Z._Z.A -0.16
#>   2: 2020-01-03 BBSIS.D.I.ZAR.ZI.EUR.S1311.B.A604.R10XX.R.A.A._Z._Z.A -0.27
#>   3: 2020-01-06 BBSIS.D.I.ZAR.ZI.EUR.S1311.B.A604.R10XX.R.A.A._Z._Z.A -0.27
#>   4: 2020-01-07 BBSIS.D.I.ZAR.ZI.EUR.S1311.B.A604.R10XX.R.A.A._Z._Z.A -0.27
#>   5: 2020-01-08 BBSIS.D.I.ZAR.ZI.EUR.S1311.B.A604.R10XX.R.A.A._Z._Z.A -0.27
#>  ---                                                                       
#> 144: 2020-07-27 BBSIS.D.I.ZAR.ZI.EUR.S1311.B.A604.R10XX.R.A.A._Z._Z.A -0.48
#> 145: 2020-07-28 BBSIS.D.I.ZAR.ZI.EUR.S1311.B.A604.R10XX.R.A.A._Z._Z.A -0.52
#> 146: 2020-07-29 BBSIS.D.I.ZAR.ZI.EUR.S1311.B.A604.R10XX.R.A.A._Z._Z.A -0.53
#> 147: 2020-07-30 BBSIS.D.I.ZAR.ZI.EUR.S1311.B.A604.R10XX.R.A.A._Z._Z.A -0.55
#> 148: 2020-07-31 BBSIS.D.I.ZAR.ZI.EUR.S1311.B.A604.R10XX.R.A.A._Z._Z.A -0.57
#>                                                                                                                       title
#>                                                                                                                      <char>
#>   1: Aus der Zinsstruktur abgeleitete Renditen für Bundeswertpapiere  mit jährl. Kuponzahlungen / RLZ 10 Jahre / Tageswerte
#>   2: Aus der Zinsstruktur abgeleitete Renditen für Bundeswertpapiere  mit jährl. Kuponzahlungen / RLZ 10 Jahre / Tageswerte
#>   3: Aus der Zinsstruktur abgeleitete Renditen für Bundeswertpapiere  mit jährl. Kuponzahlungen / RLZ 10 Jahre / Tageswerte
#>   4: Aus der Zinsstruktur abgeleitete Renditen für Bundeswertpapiere  mit jährl. Kuponzahlungen / RLZ 10 Jahre / Tageswerte
#>   5: Aus der Zinsstruktur abgeleitete Renditen für Bundeswertpapiere  mit jährl. Kuponzahlungen / RLZ 10 Jahre / Tageswerte
#>  ---                                                                                                                       
#> 144: Aus der Zinsstruktur abgeleitete Renditen für Bundeswertpapiere  mit jährl. Kuponzahlungen / RLZ 10 Jahre / Tageswerte
#> 145: Aus der Zinsstruktur abgeleitete Renditen für Bundeswertpapiere  mit jährl. Kuponzahlungen / RLZ 10 Jahre / Tageswerte
#> 146: Aus der Zinsstruktur abgeleitete Renditen für Bundeswertpapiere  mit jährl. Kuponzahlungen / RLZ 10 Jahre / Tageswerte
#> 147: Aus der Zinsstruktur abgeleitete Renditen für Bundeswertpapiere  mit jährl. Kuponzahlungen / RLZ 10 Jahre / Tageswerte
#> 148: Aus der Zinsstruktur abgeleitete Renditen für Bundeswertpapiere  mit jährl. Kuponzahlungen / RLZ 10 Jahre / Tageswerte
#>        freq bearer_reg   item valuation currency issuer_class listed_sub
#>      <char>     <char> <char>    <char>   <char>       <char>     <char>
#>   1:  daily          I    ZAR        ZI      EUR        S1311          B
#>   2:  daily          I    ZAR        ZI      EUR        S1311          B
#>   3:  daily          I    ZAR        ZI      EUR        S1311          B
#>   4:  daily          I    ZAR        ZI      EUR        S1311          B
#>   5:  daily          I    ZAR        ZI      EUR        S1311          B
#>  ---                                                                    
#> 144:  daily          I    ZAR        ZI      EUR        S1311          B
#> 145:  daily          I    ZAR        ZI      EUR        S1311          B
#> 146:  daily          I    ZAR        ZI      EUR        S1311          B
#> 147:  daily          I    ZAR        ZI      EUR        S1311          B
#> 148:  daily          I    ZAR        ZI      EUR        S1311          B
#>      security_class maturity interest_type interest_rate redemption certificate
#>              <char>   <char>        <char>        <char>     <char>      <char>
#>   1:           A604    R10XX             R             A          A          _Z
#>   2:           A604    R10XX             R             A          A          _Z
#>   3:           A604    R10XX             R             A          A          _Z
#>   4:           A604    R10XX             R             A          A          _Z
#>   5:           A604    R10XX             R             A          A          _Z
#>  ---                                                                           
#> 144:           A604    R10XX             R             A          A          _Z
#> 145:           A604    R10XX             R             A          A          _Z
#> 146:           A604    R10XX             R             A          A          _Z
#> 147:           A604    R10XX             R             A          A          _Z
#> 148:           A604    R10XX             R             A          A          _Z
#>      coverage rating time_format decimals    unit unit_mult category
#>        <char> <char>      <char>    <int>  <char>    <char>   <char>
#>   1:       _Z      A         P1D        2 PROZENT         0     GKZR
#>   2:       _Z      A         P1D        2 PROZENT         0     GKZR
#>   3:       _Z      A         P1D        2 PROZENT         0     GKZR
#>   4:       _Z      A         P1D        2 PROZENT         0     GKZR
#>   5:       _Z      A         P1D        2 PROZENT         0     GKZR
#>  ---                                                                
#> 144:       _Z      A         P1D        2 PROZENT         0     GKZR
#> 145:       _Z      A         P1D        2 PROZENT         0     GKZR
#> 146:       _Z      A         P1D        2 PROZENT         0     GKZR
#> 147:       _Z      A         P1D        2 PROZENT         0     GKZR
#> 148:       _Z      A         P1D        2 PROZENT         0     GKZR
# or only specify the start date
bbk_data(
  "BBSIS", "D.I.ZAR.ZI.EUR.S1311.B.A604.R10XX.R.A.A._Z._Z.A",
  start_period = "2024-04-01"
)
#>            date                                                   key value
#>          <Date>                                                <char> <num>
#>   1: 2024-04-02 BBSIS.D.I.ZAR.ZI.EUR.S1311.B.A604.R10XX.R.A.A._Z._Z.A  2.37
#>   2: 2024-04-03 BBSIS.D.I.ZAR.ZI.EUR.S1311.B.A604.R10XX.R.A.A._Z._Z.A  2.41
#>   3: 2024-04-04 BBSIS.D.I.ZAR.ZI.EUR.S1311.B.A604.R10XX.R.A.A._Z._Z.A  2.39
#>   4: 2024-04-05 BBSIS.D.I.ZAR.ZI.EUR.S1311.B.A604.R10XX.R.A.A._Z._Z.A  2.37
#>   5: 2024-04-08 BBSIS.D.I.ZAR.ZI.EUR.S1311.B.A604.R10XX.R.A.A._Z._Z.A  2.46
#>  ---                                                                       
#> 346: 2025-08-11 BBSIS.D.I.ZAR.ZI.EUR.S1311.B.A604.R10XX.R.A.A._Z._Z.A  2.69
#> 347: 2025-08-12 BBSIS.D.I.ZAR.ZI.EUR.S1311.B.A604.R10XX.R.A.A._Z._Z.A  2.71
#> 348: 2025-08-13 BBSIS.D.I.ZAR.ZI.EUR.S1311.B.A604.R10XX.R.A.A._Z._Z.A  2.71
#> 349: 2025-08-14 BBSIS.D.I.ZAR.ZI.EUR.S1311.B.A604.R10XX.R.A.A._Z._Z.A  2.67
#> 350: 2025-08-15 BBSIS.D.I.ZAR.ZI.EUR.S1311.B.A604.R10XX.R.A.A._Z._Z.A  2.75
#>                                                                                                                       title
#>                                                                                                                      <char>
#>   1: Aus der Zinsstruktur abgeleitete Renditen für Bundeswertpapiere  mit jährl. Kuponzahlungen / RLZ 10 Jahre / Tageswerte
#>   2: Aus der Zinsstruktur abgeleitete Renditen für Bundeswertpapiere  mit jährl. Kuponzahlungen / RLZ 10 Jahre / Tageswerte
#>   3: Aus der Zinsstruktur abgeleitete Renditen für Bundeswertpapiere  mit jährl. Kuponzahlungen / RLZ 10 Jahre / Tageswerte
#>   4: Aus der Zinsstruktur abgeleitete Renditen für Bundeswertpapiere  mit jährl. Kuponzahlungen / RLZ 10 Jahre / Tageswerte
#>   5: Aus der Zinsstruktur abgeleitete Renditen für Bundeswertpapiere  mit jährl. Kuponzahlungen / RLZ 10 Jahre / Tageswerte
#>  ---                                                                                                                       
#> 346: Aus der Zinsstruktur abgeleitete Renditen für Bundeswertpapiere  mit jährl. Kuponzahlungen / RLZ 10 Jahre / Tageswerte
#> 347: Aus der Zinsstruktur abgeleitete Renditen für Bundeswertpapiere  mit jährl. Kuponzahlungen / RLZ 10 Jahre / Tageswerte
#> 348: Aus der Zinsstruktur abgeleitete Renditen für Bundeswertpapiere  mit jährl. Kuponzahlungen / RLZ 10 Jahre / Tageswerte
#> 349: Aus der Zinsstruktur abgeleitete Renditen für Bundeswertpapiere  mit jährl. Kuponzahlungen / RLZ 10 Jahre / Tageswerte
#> 350: Aus der Zinsstruktur abgeleitete Renditen für Bundeswertpapiere  mit jährl. Kuponzahlungen / RLZ 10 Jahre / Tageswerte
#>        freq bearer_reg   item valuation currency issuer_class listed_sub
#>      <char>     <char> <char>    <char>   <char>       <char>     <char>
#>   1:  daily          I    ZAR        ZI      EUR        S1311          B
#>   2:  daily          I    ZAR        ZI      EUR        S1311          B
#>   3:  daily          I    ZAR        ZI      EUR        S1311          B
#>   4:  daily          I    ZAR        ZI      EUR        S1311          B
#>   5:  daily          I    ZAR        ZI      EUR        S1311          B
#>  ---                                                                    
#> 346:  daily          I    ZAR        ZI      EUR        S1311          B
#> 347:  daily          I    ZAR        ZI      EUR        S1311          B
#> 348:  daily          I    ZAR        ZI      EUR        S1311          B
#> 349:  daily          I    ZAR        ZI      EUR        S1311          B
#> 350:  daily          I    ZAR        ZI      EUR        S1311          B
#>      security_class maturity interest_type interest_rate redemption certificate
#>              <char>   <char>        <char>        <char>     <char>      <char>
#>   1:           A604    R10XX             R             A          A          _Z
#>   2:           A604    R10XX             R             A          A          _Z
#>   3:           A604    R10XX             R             A          A          _Z
#>   4:           A604    R10XX             R             A          A          _Z
#>   5:           A604    R10XX             R             A          A          _Z
#>  ---                                                                           
#> 346:           A604    R10XX             R             A          A          _Z
#> 347:           A604    R10XX             R             A          A          _Z
#> 348:           A604    R10XX             R             A          A          _Z
#> 349:           A604    R10XX             R             A          A          _Z
#> 350:           A604    R10XX             R             A          A          _Z
#>      coverage rating time_format decimals    unit unit_mult category
#>        <char> <char>      <char>    <int>  <char>    <char>   <char>
#>   1:       _Z      A         P1D        2 PROZENT         0     GKZR
#>   2:       _Z      A         P1D        2 PROZENT         0     GKZR
#>   3:       _Z      A         P1D        2 PROZENT         0     GKZR
#>   4:       _Z      A         P1D        2 PROZENT         0     GKZR
#>   5:       _Z      A         P1D        2 PROZENT         0     GKZR
#>  ---                                                                
#> 346:       _Z      A         P1D        2 PROZENT         0     GKZR
#> 347:       _Z      A         P1D        2 PROZENT         0     GKZR
#> 348:       _Z      A         P1D        2 PROZENT         0     GKZR
#> 349:       _Z      A         P1D        2 PROZENT         0     GKZR
#> 350:       _Z      A         P1D        2 PROZENT         0     GKZR
# }