Returns the available metadata

Retrieval of the metadata stored in the Bundesbank's time series database. Access via the SDMX Web Service API of the Bundesbank.

Usage

bbk_metadata(type, id = NULL, lang = c("en", "de"))

Source

https://www.bundesbank.de/en/statistics/time-series-databases/help-for-sdmx-web-service/web-service-interface-metadata

Arguments

type

character(1) the type of metadata to query. One of: "datastructure", "dataflow", "codelist", or "concept".

id

character(1) id to query. Default NULL.

lang

character(1) language to query, either "en" or "de". Default "en".

Value

A data.frame() with the queried metadata. The columns are:

id

The id of the metadata

name

The name of the metadata

See also

Examples

# \donttest{
bbk_metadata("datastructure")
#> # A tibble: 73 × 2
#>    id        name                                                               
#>    <chr>     <chr>                                                              
#>  1 BBK_BSPL  Profit and loss                                                    
#>  2 BBK_DOES  Short-term economic indicators (classified by economic sector)     
#>  3 BBK_ERX   ECB euro reference exchange rates, exchange rates in particular co…
#>  4 BBK_ESBOP External sector Balance of Payments according to BPM6              
#>  5 BBK_EXBS  Business and consumer surveys of foreign countries                 
#>  6 BBK_DOSC  System of  indicators for the German commercial real estate market 
#>  7 BBK_DOSR  System of indicators for the German residential property market    
#>  8 BBK_EXNA  National Accounts of foreign countries                             
#>  9 BBK_EXPR  Consumer Prices of foreign countries                               
#> 10 BBK_EXEI  Leading indicators of foreign countries                            
#> # ℹ 63 more rows
bbk_metadata("dataflow", "BBSIS")
#> # A tibble: 1 × 2
#>   id    name                                
#>   <chr> <chr>                               
#> 1 BBSIS Securities Issues Statistics (Bonds)
bbk_metadata("codelist", "CL_BBK_ACIP_ASSET_LIABILITY")
#> # A tibble: 3 × 2
#>   id                          name                
#>   <chr>                       <chr>               
#> 1 CL_BBK_ACIP_ASSET_LIABILITY Assets / Liabilities
#> 2 CL_BBK_ACIP_ASSET_LIABILITY Assets              
#> 3 CL_BBK_ACIP_ASSET_LIABILITY Liabilities         
bbk_metadata("concept", "CS_BBK_BSPL")
#> # A tibble: 38 × 2
#>    id          name                                         
#>    <chr>       <chr>                                        
#>  1 CS_BBK_BSPL Profit and loss                              
#>  2 CS_BBK_BSPL Frequency (BBk)                              
#>  3 CS_BBK_BSPL Area (countries, list of countries), BBk only
#>  4 CS_BBK_BSPL Adjustment (BBk)                             
#>  5 CS_BBK_BSPL Reference sector breakdown                   
#>  6 CS_BBK_BSPL Profit/Loss Item                             
#>  7 CS_BBK_BSPL Original maturity                            
#>  8 CS_BBK_BSPL Data type                                    
#>  9 CS_BBK_BSPL Counterpart area                             
#> 10 CS_BBK_BSPL Counterpart Sector                           
#> # ℹ 28 more rows
# }