Returns the Bundesbank time serie that is found with the specified time series key

Returns the Bundesbank time serie that is found with the specified time series key

Usage

bbk_series(key)

Source

https://www.bundesbank.de/en/statistics/time-series-databases/help-for-sdmx-web-service/web-service-interface-data

Arguments

key

(character(1)) key to query.

Value

A data.table::data.table() with the requested data.

See also

Examples

# \donttest{
bbk_series("BBEX3.M.DKK.EUR.BB.AC.A01")
#>            date                       key  value
#>          <Date>                    <char>  <num>
#>   1: 1999-02-01 BBEX3.M.DKK.EUR.BB.AC.A01 7.4344
#>   2: 1999-03-01 BBEX3.M.DKK.EUR.BB.AC.A01 7.4317
#>   3: 1999-04-01 BBEX3.M.DKK.EUR.BB.AC.A01 7.4323
#>   4: 1999-05-01 BBEX3.M.DKK.EUR.BB.AC.A01 7.4318
#>   5: 1999-06-01 BBEX3.M.DKK.EUR.BB.AC.A01 7.4341
#>  ---                                            
#> 312: 2025-01-01 BBEX3.M.DKK.EUR.BB.AC.A01 7.4618
#> 313: 2025-02-01 BBEX3.M.DKK.EUR.BB.AC.A01 7.4583
#> 314: 2025-03-01 BBEX3.M.DKK.EUR.BB.AC.A01 7.4613
#> 315: 2025-04-01 BBEX3.M.DKK.EUR.BB.AC.A01 7.4636
#> 316: 2025-05-01 BBEX3.M.DKK.EUR.BB.AC.A01 7.4587
#>                                                                            title
#>                                                                           <char>
#>   1: Euro foreign exchange reference rate of the ECB / EUR 1 = DKK ... / Denmark
#>   2: Euro foreign exchange reference rate of the ECB / EUR 1 = DKK ... / Denmark
#>   3: Euro foreign exchange reference rate of the ECB / EUR 1 = DKK ... / Denmark
#>   4: Euro foreign exchange reference rate of the ECB / EUR 1 = DKK ... / Denmark
#>   5: Euro foreign exchange reference rate of the ECB / EUR 1 = DKK ... / Denmark
#>  ---                                                                            
#> 312: Euro foreign exchange reference rate of the ECB / EUR 1 = DKK ... / Denmark
#> 313: Euro foreign exchange reference rate of the ECB / EUR 1 = DKK ... / Denmark
#> 314: Euro foreign exchange reference rate of the ECB / EUR 1 = DKK ... / Denmark
#> 315: Euro foreign exchange reference rate of the ECB / EUR 1 = DKK ... / Denmark
#> 316: Euro foreign exchange reference rate of the ECB / EUR 1 = DKK ... / Denmark
#>         freq category   unit unit_mult         last_update
#>       <char>   <char> <char>    <char>              <char>
#>   1: monthly     WEDE    DKK       One 2025-05-30 16:04:04
#>   2: monthly     WEDE    DKK       One 2025-05-30 16:04:04
#>   3: monthly     WEDE    DKK       One 2025-05-30 16:04:04
#>   4: monthly     WEDE    DKK       One 2025-05-30 16:04:04
#>   5: monthly     WEDE    DKK       One 2025-05-30 16:04:04
#>  ---                                                      
#> 312: monthly     WEDE    DKK       One 2025-05-30 16:04:04
#> 313: monthly     WEDE    DKK       One 2025-05-30 16:04:04
#> 314: monthly     WEDE    DKK       One 2025-05-30 16:04:04
#> 315: monthly     WEDE    DKK       One 2025-05-30 16:04:04
#> 316: monthly     WEDE    DKK       One 2025-05-30 16:04:04
#>                                                            comment
#>                                                             <char>
#>   1: The ECB publishes daily euro foreign exchange reference rates
#>   2: The ECB publishes daily euro foreign exchange reference rates
#>   3: The ECB publishes daily euro foreign exchange reference rates
#>   4: The ECB publishes daily euro foreign exchange reference rates
#>   5: The ECB publishes daily euro foreign exchange reference rates
#>  ---                                                              
#> 312: The ECB publishes daily euro foreign exchange reference rates
#> 313: The ECB publishes daily euro foreign exchange reference rates
#> 314: The ECB publishes daily euro foreign exchange reference rates
#> 315: The ECB publishes daily euro foreign exchange reference rates
#> 316: The ECB publishes daily euro foreign exchange reference rates
#>                            source
#>                            <char>
#>   1: European Central Bank (ECB).
#>   2: European Central Bank (ECB).
#>   3: European Central Bank (ECB).
#>   4: European Central Bank (ECB).
#>   5: European Central Bank (ECB).
#>  ---                             
#> 312: European Central Bank (ECB).
#> 313: European Central Bank (ECB).
#> 314: European Central Bank (ECB).
#> 315: European Central Bank (ECB).
#> 316: European Central Bank (ECB).
bbk_series("BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B")
#>         date                                 key value
#>       <char>                              <char> <num>
#>   1: 1999-Q4 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.4
#>   2: 2000-Q1 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.4
#>   3: 2000-Q2 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.4
#>   4: 2000-Q3 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.4
#>   5: 2000-Q4 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.4
#>  ---                                                  
#>  97: 2023-Q4 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.7
#>  98: 2024-Q1 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.8
#>  99: 2024-Q2 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   5.3
#> 100: 2024-Q3 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   5.0
#> 101: 2024-Q4 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.3
#>                                                                                       title
#>                                                                                      <char>
#>   1: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>   2: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>   3: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>   4: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>   5: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  ---                                                                                       
#>  97: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  98: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  99: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#> 100: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#> 101: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>           freq category   unit unit_mult         last_update comment source
#>         <char>   <char> <char>    <char>              <char>  <char> <char>
#>   1: quarterly     GRF2   Euro  Billions 2025-04-24 09:43:46    <NA>   <NA>
#>   2: quarterly     GRF2   Euro  Billions 2025-04-24 09:43:46    <NA>   <NA>
#>   3: quarterly     GRF2   Euro  Billions 2025-04-24 09:43:46    <NA>   <NA>
#>   4: quarterly     GRF2   Euro  Billions 2025-04-24 09:43:46    <NA>   <NA>
#>   5: quarterly     GRF2   Euro  Billions 2025-04-24 09:43:46    <NA>   <NA>
#>  ---                                                                       
#>  97: quarterly     GRF2   Euro  Billions 2025-04-24 09:43:46    <NA>   <NA>
#>  98: quarterly     GRF2   Euro  Billions 2025-04-24 09:43:46    <NA>   <NA>
#>  99: quarterly     GRF2   Euro  Billions 2025-04-24 09:43:46    <NA>   <NA>
#> 100: quarterly     GRF2   Euro  Billions 2025-04-24 09:43:46    <NA>   <NA>
#> 101: quarterly     GRF2   Euro  Billions 2025-04-24 09:43:46    <NA>   <NA>
bbk_series("BBBK11.D.TTA000")
#>             date             key  value
#>           <Date>          <char> <char>
#>    1: 1999-01-01 BBBK11.D.TTA000  29312
#>    2: 1999-01-08 BBBK11.D.TTA000  27475
#>    3: 1999-01-15 BBBK11.D.TTA000  27475
#>    4: 1999-01-22 BBBK11.D.TTA000  27475
#>    5: 1999-01-29 BBBK11.D.TTA000  27475
#>   ---                                  
#> 1377: 2025-05-09 BBBK11.D.TTA000 310885
#> 1378: 2025-05-16 BBBK11.D.TTA000 310885
#> 1379: 2025-05-23 BBBK11.D.TTA000 310806
#> 1380: 2025-05-30 BBBK11.D.TTA000 310806
#> 1381: 2025-06-06 BBBK11.D.TTA000 310806
#>                                                              title   freq
#>                                                             <char> <char>
#>    1: Gold and gold receivables / unadjusted / Deutsche Bundesbank  daily
#>    2: Gold and gold receivables / unadjusted / Deutsche Bundesbank  daily
#>    3: Gold and gold receivables / unadjusted / Deutsche Bundesbank  daily
#>    4: Gold and gold receivables / unadjusted / Deutsche Bundesbank  daily
#>    5: Gold and gold receivables / unadjusted / Deutsche Bundesbank  daily
#>   ---                                                                    
#> 1377: Gold and gold receivables / unadjusted / Deutsche Bundesbank  daily
#> 1378: Gold and gold receivables / unadjusted / Deutsche Bundesbank  daily
#> 1379: Gold and gold receivables / unadjusted / Deutsche Bundesbank  daily
#> 1380: Gold and gold receivables / unadjusted / Deutsche Bundesbank  daily
#> 1381: Gold and gold receivables / unadjusted / Deutsche Bundesbank  daily
#>       category   unit unit_mult         last_update comment source
#>         <char> <char>    <char>              <char>  <char> <char>
#>    1:     BABA   EURO  Millions 2025-06-12 15:13:12    <NA>   <NA>
#>    2:     BABA   EURO  Millions 2025-06-12 15:13:12    <NA>   <NA>
#>    3:     BABA   EURO  Millions 2025-06-12 15:13:12    <NA>   <NA>
#>    4:     BABA   EURO  Millions 2025-06-12 15:13:12    <NA>   <NA>
#>    5:     BABA   EURO  Millions 2025-06-12 15:13:12    <NA>   <NA>
#>   ---                                                             
#> 1377:     BABA   EURO  Millions 2025-06-12 15:13:12    <NA>   <NA>
#> 1378:     BABA   EURO  Millions 2025-06-12 15:13:12    <NA>   <NA>
#> 1379:     BABA   EURO  Millions 2025-06-12 15:13:12    <NA>   <NA>
#> 1380:     BABA   EURO  Millions 2025-06-12 15:13:12    <NA>   <NA>
#> 1381:     BABA   EURO  Millions 2025-06-12 15:13:12    <NA>   <NA>
# }