Returns the Bundesbank time serie that is found with the specified time series key

Returns the Bundesbank time serie that is found with the specified time series key

Usage

bbk_series(key)

Source

https://www.bundesbank.de/en/statistics/time-series-databases/help-for-sdmx-web-service/web-service-interface-data

Arguments

key

(character(1)) key to query.

Value

A data.table::data.table() with the requested data.

See also

Examples

# \donttest{
bbk_series("BBEX3.M.DKK.EUR.BB.AC.A01")
#>            date                       key  value
#>          <Date>                    <char>  <num>
#>   1: 1999-02-01 BBEX3.M.DKK.EUR.BB.AC.A01 7.4344
#>   2: 1999-03-01 BBEX3.M.DKK.EUR.BB.AC.A01 7.4317
#>   3: 1999-04-01 BBEX3.M.DKK.EUR.BB.AC.A01 7.4323
#>   4: 1999-05-01 BBEX3.M.DKK.EUR.BB.AC.A01 7.4318
#>   5: 1999-06-01 BBEX3.M.DKK.EUR.BB.AC.A01 7.4341
#>  ---                                            
#> 307: 2024-08-01 BBEX3.M.DKK.EUR.BB.AC.A01 7.4589
#> 308: 2024-09-01 BBEX3.M.DKK.EUR.BB.AC.A01 7.4560
#> 309: 2024-10-01 BBEX3.M.DKK.EUR.BB.AC.A01 7.4601
#> 310: 2024-11-01 BBEX3.M.DKK.EUR.BB.AC.A01 7.4579
#> 311: 2024-12-01 BBEX3.M.DKK.EUR.BB.AC.A01 7.4578
#>                                                                            title
#>                                                                           <char>
#>   1: Euro foreign exchange reference rate of the ECB / EUR 1 = DKK ... / Denmark
#>   2: Euro foreign exchange reference rate of the ECB / EUR 1 = DKK ... / Denmark
#>   3: Euro foreign exchange reference rate of the ECB / EUR 1 = DKK ... / Denmark
#>   4: Euro foreign exchange reference rate of the ECB / EUR 1 = DKK ... / Denmark
#>   5: Euro foreign exchange reference rate of the ECB / EUR 1 = DKK ... / Denmark
#>  ---                                                                            
#> 307: Euro foreign exchange reference rate of the ECB / EUR 1 = DKK ... / Denmark
#> 308: Euro foreign exchange reference rate of the ECB / EUR 1 = DKK ... / Denmark
#> 309: Euro foreign exchange reference rate of the ECB / EUR 1 = DKK ... / Denmark
#> 310: Euro foreign exchange reference rate of the ECB / EUR 1 = DKK ... / Denmark
#> 311: Euro foreign exchange reference rate of the ECB / EUR 1 = DKK ... / Denmark
#>         freq category   unit unit_mult         last_update
#>       <char>   <char> <char>    <char>              <char>
#>   1: monthly     WEDE    DKK       One 2024-12-31 16:00:09
#>   2: monthly     WEDE    DKK       One 2024-12-31 16:00:09
#>   3: monthly     WEDE    DKK       One 2024-12-31 16:00:09
#>   4: monthly     WEDE    DKK       One 2024-12-31 16:00:09
#>   5: monthly     WEDE    DKK       One 2024-12-31 16:00:09
#>  ---                                                      
#> 307: monthly     WEDE    DKK       One 2024-12-31 16:00:09
#> 308: monthly     WEDE    DKK       One 2024-12-31 16:00:09
#> 309: monthly     WEDE    DKK       One 2024-12-31 16:00:09
#> 310: monthly     WEDE    DKK       One 2024-12-31 16:00:09
#> 311: monthly     WEDE    DKK       One 2024-12-31 16:00:09
#>                                                            comment
#>                                                             <char>
#>   1: The ECB publishes daily euro foreign exchange reference rates
#>   2: The ECB publishes daily euro foreign exchange reference rates
#>   3: The ECB publishes daily euro foreign exchange reference rates
#>   4: The ECB publishes daily euro foreign exchange reference rates
#>   5: The ECB publishes daily euro foreign exchange reference rates
#>  ---                                                              
#> 307: The ECB publishes daily euro foreign exchange reference rates
#> 308: The ECB publishes daily euro foreign exchange reference rates
#> 309: The ECB publishes daily euro foreign exchange reference rates
#> 310: The ECB publishes daily euro foreign exchange reference rates
#> 311: The ECB publishes daily euro foreign exchange reference rates
#>                            source
#>                            <char>
#>   1: European Central Bank (ECB).
#>   2: European Central Bank (ECB).
#>   3: European Central Bank (ECB).
#>   4: European Central Bank (ECB).
#>   5: European Central Bank (ECB).
#>  ---                             
#> 307: European Central Bank (ECB).
#> 308: European Central Bank (ECB).
#> 309: European Central Bank (ECB).
#> 310: European Central Bank (ECB).
#> 311: European Central Bank (ECB).
bbk_series("BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B")
#>         date                                 key value
#>       <char>                              <char> <num>
#>   1: 1999-Q4 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.4
#>   2: 2000-Q1 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.4
#>   3: 2000-Q2 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.4
#>   4: 2000-Q3 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.4
#>   5: 2000-Q4 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.4
#>   6: 2001-Q1 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.4
#>   7: 2001-Q2 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.4
#>   8: 2001-Q3 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.4
#>   9: 2001-Q4 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.4
#>  10: 2002-Q1 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.4
#>  11: 2002-Q2 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.4
#>  12: 2002-Q3 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.4
#>  13: 2002-Q4 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.4
#>  14: 2003-Q1 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.4
#>  15: 2003-Q2 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.4
#>  16: 2003-Q3 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.4
#>  17: 2003-Q4 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.4
#>  18: 2004-Q1 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.4
#>  19: 2004-Q2 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.4
#>  20: 2004-Q3 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.4
#>  21: 2004-Q4 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.4
#>  22: 2005-Q1 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.4
#>  23: 2005-Q2 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.4
#>  24: 2005-Q3 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.4
#>  25: 2005-Q4 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.4
#>  26: 2006-Q1 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.4
#>  27: 2006-Q2 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.4
#>  28: 2006-Q3 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.4
#>  29: 2006-Q4 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.4
#>  30: 2007-Q1 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.4
#>  31: 2007-Q2 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.4
#>  32: 2007-Q3 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.4
#>  33: 2007-Q4 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.4
#>  34: 2008-Q1 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.4
#>  35: 2008-Q2 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.4
#>  36: 2008-Q3 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.4
#>  37: 2008-Q4 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.4
#>  38: 2009-Q1 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.4
#>  39: 2009-Q2 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.4
#>  40: 2009-Q3 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.4
#>  41: 2009-Q4 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.4
#>  42: 2010-Q1 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.4
#>  43: 2010-Q2 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.4
#>  44: 2010-Q3 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.4
#>  45: 2010-Q4 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.4
#>  46: 2011-Q1 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.4
#>  47: 2011-Q2 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.4
#>  48: 2011-Q3 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.4
#>  49: 2011-Q4 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.4
#>  50: 2012-Q1 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.4
#>  51: 2012-Q2 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.4
#>  52: 2012-Q3 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.4
#>  53: 2012-Q4 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.4
#>  54: 2013-Q1 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.7
#>  55: 2013-Q2 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.7
#>  56: 2013-Q3 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.7
#>  57: 2013-Q4 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.7
#>  58: 2014-Q1 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.7
#>  59: 2014-Q2 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.7
#>  60: 2014-Q3 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.7
#>  61: 2014-Q4 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.7
#>  62: 2015-Q1 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.6
#>  63: 2015-Q2 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.6
#>  64: 2015-Q3 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.6
#>  65: 2015-Q4 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.6
#>  66: 2016-Q1 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.6
#>  67: 2016-Q2 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.6
#>  68: 2016-Q3 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.6
#>  69: 2016-Q4 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.6
#>  70: 2017-Q1 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.6
#>  71: 2017-Q2 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.6
#>  72: 2017-Q3 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.6
#>  73: 2017-Q4 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.6
#>  74: 2018-Q1 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.6
#>  75: 2018-Q2 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.6
#>  76: 2018-Q3 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.6
#>  77: 2018-Q4 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.6
#>  78: 2019-Q1 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.6
#>  79: 2019-Q2 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.6
#>  80: 2019-Q3 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.6
#>  81: 2019-Q4 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.6
#>  82: 2020-Q1 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.6
#>  83: 2020-Q2 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.6
#>  84: 2020-Q3 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.6
#>  85: 2020-Q4 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   5.5
#>  86: 2021-Q1 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   5.7
#>  87: 2021-Q2 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   5.5
#>  88: 2021-Q3 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   5.3
#>  89: 2021-Q4 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   5.3
#>  90: 2022-Q1 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   5.3
#>  91: 2022-Q2 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.9
#>  92: 2022-Q3 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.8
#>  93: 2022-Q4 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.6
#>  94: 2023-Q1 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.6
#>  95: 2023-Q2 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   5.1
#>  96: 2023-Q3 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   5.3
#>  97: 2023-Q4 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.7
#>  98: 2024-Q1 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   4.8
#>  99: 2024-Q2 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   5.3
#> 100: 2024-Q3 BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B   5.0
#>         date                                 key value
#>                                                                                       title
#>                                                                                      <char>
#>   1: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>   2: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>   3: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>   4: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>   5: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>   6: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>   7: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>   8: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>   9: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  10: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  11: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  12: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  13: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  14: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  15: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  16: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  17: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  18: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  19: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  20: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  21: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  22: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  23: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  24: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  25: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  26: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  27: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  28: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  29: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  30: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  31: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  32: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  33: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  34: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  35: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  36: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  37: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  38: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  39: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  40: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  41: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  42: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  43: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  44: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  45: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  46: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  47: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  48: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  49: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  50: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  51: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  52: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  53: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  54: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  55: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  56: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  57: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  58: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  59: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  60: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  61: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  62: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  63: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  64: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  65: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  66: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  67: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  68: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  69: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  70: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  71: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  72: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  73: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  74: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  75: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  76: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  77: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  78: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  79: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  80: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  81: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  82: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  83: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  84: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  85: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  86: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  87: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  88: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  89: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  90: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  91: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  92: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  93: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  94: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  95: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  96: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  97: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  98: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>  99: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#> 100: Short-term loans Creditor: Central bank (Germany) / Debtor: All sectors (World) Stocks
#>                                                                                       title
#>           freq category   unit unit_mult         last_update comment source
#>         <char>   <char> <char>    <char>              <char>  <char> <char>
#>   1: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>   2: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>   3: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>   4: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>   5: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>   6: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>   7: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>   8: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>   9: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  10: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  11: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  12: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  13: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  14: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  15: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  16: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  17: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  18: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  19: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  20: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  21: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  22: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  23: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  24: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  25: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  26: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  27: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  28: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  29: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  30: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  31: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  32: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  33: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  34: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  35: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  36: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  37: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  38: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  39: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  40: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  41: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  42: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  43: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  44: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  45: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  46: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  47: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  48: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  49: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  50: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  51: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  52: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  53: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  54: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  55: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  56: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  57: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  58: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  59: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  60: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  61: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  62: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  63: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  64: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  65: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  66: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  67: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  68: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  69: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  70: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  71: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  72: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  73: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  74: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  75: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  76: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  77: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  78: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  79: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  80: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  81: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  82: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  83: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  84: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  85: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  86: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  87: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  88: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  89: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  90: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  91: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  92: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  93: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  94: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  95: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  96: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  97: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  98: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>  99: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#> 100: quarterly     GRF2   Euro  Billions 2025-01-15 09:38:57    <NA>   <NA>
#>           freq category   unit unit_mult         last_update comment source
bbk_series("BBBK11.D.TTA000")
#>             date             key  value
#>           <Date>          <char> <char>
#>    1: 1999-01-01 BBBK11.D.TTA000  29312
#>    2: 1999-01-08 BBBK11.D.TTA000  27475
#>    3: 1999-01-15 BBBK11.D.TTA000  27475
#>    4: 1999-01-22 BBBK11.D.TTA000  27475
#>    5: 1999-01-29 BBBK11.D.TTA000  27475
#>   ---                                  
#> 1356: 2024-12-13 BBBK11.D.TTA000 254267
#> 1357: 2024-12-20 BBBK11.D.TTA000 254267
#> 1358: 2024-12-27 BBBK11.D.TTA000 254267
#> 1359: 2025-01-03 BBBK11.D.TTA000 270580
#> 1360: 2025-01-10 BBBK11.D.TTA000 270580
#>                                                              title   freq
#>                                                             <char> <char>
#>    1: Gold and gold receivables / unadjusted / Deutsche Bundesbank  daily
#>    2: Gold and gold receivables / unadjusted / Deutsche Bundesbank  daily
#>    3: Gold and gold receivables / unadjusted / Deutsche Bundesbank  daily
#>    4: Gold and gold receivables / unadjusted / Deutsche Bundesbank  daily
#>    5: Gold and gold receivables / unadjusted / Deutsche Bundesbank  daily
#>   ---                                                                    
#> 1356: Gold and gold receivables / unadjusted / Deutsche Bundesbank  daily
#> 1357: Gold and gold receivables / unadjusted / Deutsche Bundesbank  daily
#> 1358: Gold and gold receivables / unadjusted / Deutsche Bundesbank  daily
#> 1359: Gold and gold receivables / unadjusted / Deutsche Bundesbank  daily
#> 1360: Gold and gold receivables / unadjusted / Deutsche Bundesbank  daily
#>       category   unit unit_mult         last_update comment source
#>         <char> <char>    <char>              <char>  <char> <char>
#>    1:     BABA   EURO  Millions 2025-01-14 15:24:23    <NA>   <NA>
#>    2:     BABA   EURO  Millions 2025-01-14 15:24:23    <NA>   <NA>
#>    3:     BABA   EURO  Millions 2025-01-14 15:24:23    <NA>   <NA>
#>    4:     BABA   EURO  Millions 2025-01-14 15:24:23    <NA>   <NA>
#>    5:     BABA   EURO  Millions 2025-01-14 15:24:23    <NA>   <NA>
#>   ---                                                             
#> 1356:     BABA   EURO  Millions 2025-01-14 15:24:23    <NA>   <NA>
#> 1357:     BABA   EURO  Millions 2025-01-14 15:24:23    <NA>   <NA>
#> 1358:     BABA   EURO  Millions 2025-01-14 15:24:23    <NA>   <NA>
#> 1359:     BABA   EURO  Millions 2025-01-14 15:24:23    <NA>   <NA>
#> 1360:     BABA   EURO  Millions 2025-01-14 15:24:23    <NA>   <NA>
# }