Returns the Bundesbank time serie that is found with the specified time series key

Returns the Bundesbank time serie that is found with the specified time series key

Usage

bbk_series(key)

Source

https://www.bundesbank.de/en/statistics/time-series-databases/help-for-sdmx-web-service/web-service-interface-data

Arguments

key

character(1) key to query.

Value

A data.frame() with the requested data.

See also

Examples

# \donttest{
bbk_series("BBEX3.M.DKK.EUR.BB.AC.A01")
#> # A tibble: 307 × 11
#>    date       key         value title freq  category unit  unit_mult last_update
#>    <date>     <chr>       <chr> <chr> <chr> <chr>    <chr> <chr>     <chr>      
#>  1 1999-02-01 BBEX3.M.DK… 7.43… Euro… mont… WEDE     DKK   One       2024-08-30…
#>  2 1999-03-01 BBEX3.M.DK… 7.43… Euro… mont… WEDE     DKK   One       2024-08-30…
#>  3 1999-04-01 BBEX3.M.DK… 7.43… Euro… mont… WEDE     DKK   One       2024-08-30…
#>  4 1999-05-01 BBEX3.M.DK… 7.43… Euro… mont… WEDE     DKK   One       2024-08-30…
#>  5 1999-06-01 BBEX3.M.DK… 7.43… Euro… mont… WEDE     DKK   One       2024-08-30…
#>  6 1999-07-01 BBEX3.M.DK… 7.44… Euro… mont… WEDE     DKK   One       2024-08-30…
#>  7 1999-08-01 BBEX3.M.DK… 7.43… Euro… mont… WEDE     DKK   One       2024-08-30…
#>  8 1999-09-01 BBEX3.M.DK… 7.43… Euro… mont… WEDE     DKK   One       2024-08-30…
#>  9 1999-10-01 BBEX3.M.DK… 7.43… Euro… mont… WEDE     DKK   One       2024-08-30…
#> 10 1999-11-01 BBEX3.M.DK… 7.43… Euro… mont… WEDE     DKK   One       2024-08-30…
#> # ℹ 297 more rows
#> # ℹ 2 more variables: comment <chr>, source <chr>
bbk_series("BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B")
#> # A tibble: 98 × 11
#>    date    key    value title freq  category unit  unit_mult last_update comment
#>    <chr>   <chr>  <chr> <chr> <chr> <chr>    <chr> <chr>     <chr>       <chr>  
#>  1 1999-Q4 BBAF3… 4.4   Shor… quar… GRF2     Euro  Billions  2024-07-17… NA     
#>  2 2000-Q1 BBAF3… 4.4   Shor… quar… GRF2     Euro  Billions  2024-07-17… NA     
#>  3 2000-Q2 BBAF3… 4.4   Shor… quar… GRF2     Euro  Billions  2024-07-17… NA     
#>  4 2000-Q3 BBAF3… 4.4   Shor… quar… GRF2     Euro  Billions  2024-07-17… NA     
#>  5 2000-Q4 BBAF3… 4.4   Shor… quar… GRF2     Euro  Billions  2024-07-17… NA     
#>  6 2001-Q1 BBAF3… 4.4   Shor… quar… GRF2     Euro  Billions  2024-07-17… NA     
#>  7 2001-Q2 BBAF3… 4.4   Shor… quar… GRF2     Euro  Billions  2024-07-17… NA     
#>  8 2001-Q3 BBAF3… 4.4   Shor… quar… GRF2     Euro  Billions  2024-07-17… NA     
#>  9 2001-Q4 BBAF3… 4.4   Shor… quar… GRF2     Euro  Billions  2024-07-17… NA     
#> 10 2002-Q1 BBAF3… 4.4   Shor… quar… GRF2     Euro  Billions  2024-07-17… NA     
#> # ℹ 88 more rows
#> # ℹ 1 more variable: source <chr>
bbk_series("BBBK11.D.TTA000")
#> # A tibble: 1,342 × 11
#>    date       key         value title freq  category unit  unit_mult last_update
#>    <date>     <chr>       <chr> <chr> <chr> <chr>    <chr> <chr>     <chr>      
#>  1 1999-01-01 BBBK11.D.T… 29312 Gold… daily BABA     EURO  Millions  2024-09-10…
#>  2 1999-01-08 BBBK11.D.T… 27475 Gold… daily BABA     EURO  Millions  2024-09-10…
#>  3 1999-01-15 BBBK11.D.T… 27475 Gold… daily BABA     EURO  Millions  2024-09-10…
#>  4 1999-01-22 BBBK11.D.T… 27475 Gold… daily BABA     EURO  Millions  2024-09-10…
#>  5 1999-01-29 BBBK11.D.T… 27475 Gold… daily BABA     EURO  Millions  2024-09-10…
#>  6 1999-02-05 BBBK11.D.T… 27475 Gold… daily BABA     EURO  Millions  2024-09-10…
#>  7 1999-02-12 BBBK11.D.T… 27475 Gold… daily BABA     EURO  Millions  2024-09-10…
#>  8 1999-02-19 BBBK11.D.T… 27475 Gold… daily BABA     EURO  Millions  2024-09-10…
#>  9 1999-02-26 BBBK11.D.T… 27475 Gold… daily BABA     EURO  Millions  2024-09-10…
#> 10 1999-03-05 BBBK11.D.T… 27475 Gold… daily BABA     EURO  Millions  2024-09-10…
#> # ℹ 1,332 more rows
#> # ℹ 2 more variables: comment <chr>, source <chr>
# }