Fetch Czech National Bank (CNB) exchange rates

Retrieve the central bank exchange rate fixing (Czech koruna against foreign currencies) from the CNB API.

Usage

cnb_fx_rates(date = NULL, year = NULL, lang = "EN")

Source

https://api.cnb.cz/cnbapi/swagger-ui.html

Arguments

date

(NULL | character(1) | Date(1))
The date to query, returning rates for all currencies on that day. If NULL, the latest available fixing is returned. Mutually exclusive with year. Default NULL.

year

(NULL | integer(1))
A calendar year, returning rates for all currencies on every working day of that year. Mutually exclusive with date. Default NULL.

lang

(character(1))
Language for the country and currency names, either "EN" or "CZ". Default "EN".

Value

A data.table::data.table() with the requested exchange rates. The rate is the amount of Czech koruna per amount units of the foreign currency.

See also

Examples

# \donttest{
# latest fixing for all currencies
cnb_fx_rates()
#>           date currency_code   currency        country amount   rate
#>         <Date>        <char>     <char>         <char>  <int>  <num>
#>  1: 2026-06-11           AUD     dollar      Australia      1 14.675
#>  2: 2026-06-11           BRL       real         Brazil      1  4.064
#>  3: 2026-06-11           CAD     dollar         Canada      1 15.003
#>  4: 2026-06-11           CNY   renminbi          China      1  3.095
#>  5: 2026-06-11           DKK      krone        Denmark      1  3.237
#>  6: 2026-06-11           EUR       euro            EMU      1 24.195
#>  7: 2026-06-11           HKD     dollar       Hongkong      1  2.677
#>  8: 2026-06-11           HUF     forint        Hungary    100  6.810
#>  9: 2026-06-11           ISK      krona        Iceland    100 16.825
#> 10: 2026-06-11           XDR        SDR            IMF      1 28.616
#> 11: 2026-06-11           INR      rupee          India    100 21.904
#> 12: 2026-06-11           IDR     rupiah      Indonesia   1000  1.167
#> 13: 2026-06-11           ILS new shekel         Israel      1  7.093
#> 14: 2026-06-11           JPY        yen          Japan    100 13.067
#> 15: 2026-06-11           MYR    ringgit       Malaysia      1  5.158
#> 16: 2026-06-11           MXN       peso         Mexico      1  1.206
#> 17: 2026-06-11           NZD     dollar    New Zealand      1 12.135
#> 18: 2026-06-11           NOK      krone         Norway      1  2.202
#> 19: 2026-06-11           PHP       peso    Philippines    100 34.200
#> 20: 2026-06-11           PLN      zloty         Poland      1  5.690
#> 21: 2026-06-11           RON        leu        Romania      1  4.618
#> 22: 2026-06-11           SGD     dollar      Singapore      1 16.290
#> 23: 2026-06-11           ZAR       rand   South Africa      1  1.271
#> 24: 2026-06-11           KRW        won    South Korea    100  1.371
#> 25: 2026-06-11           SEK      krona         Sweden      1  2.201
#> 26: 2026-06-11           CHF      franc    Switzerland      1 26.233
#> 27: 2026-06-11           THB       baht       Thailand    100 63.653
#> 28: 2026-06-11           TRY       lira         Turkey    100 45.451
#> 29: 2026-06-11           GBP      pound United Kingdom      1 28.032
#> 30: 2026-06-11           USD     dollar            USA      1 20.976
#>           date currency_code   currency        country amount   rate
#>         <Date>        <char>     <char>         <char>  <int>  <num>

# all fixings for a given year
cnb_fx_rates(year = 2024L)
#>             date currency_code amount   rate
#>           <Date>        <char>  <int>  <num>
#>    1: 2024-01-02           AUD      1 15.278
#>    2: 2024-01-02           BGN      1 12.621
#>    3: 2024-01-02           BRL      1  4.609
#>    4: 2024-01-02           CAD      1 16.938
#>    5: 2024-01-02           CHF      1 26.524
#>   ---                                       
#> 7808: 2024-12-31           THB    100 70.618
#> 7809: 2024-12-31           TRY    100 68.539
#> 7810: 2024-12-31           USD      1 24.237
#> 7811: 2024-12-31           XDR      1 31.609
#> 7812: 2024-12-31           ZAR      1  1.284
# }