Returns ECB data for a given flow and key

Returns ECB data for a given flow and key

Usage

ecb_data(
  flow,
  key = NULL,
  start_period = NULL,
  end_period = NULL,
  first_n = NULL,
  last_n = NULL
)

Source

https://data.ecb.europa.eu/help/api/data

Arguments

flow

character(1) flow to query.

key

character(1) key to query.

start_period

character(1) start date of the data. Supported formats:

YYYY for annual data (e.g., "2019")

YYYY-S[1-2] for semi-annual data (e.g., "2019-S1")

YYYY-Q[1-4] for quarterly data (e.g., "2019-Q1")

YYYY-MM for monthly data (e.g., "2019-01")

YYYY-W[01-53] for weekly data (e.g., "2019-W01")

YYYY-MM-DD for daily and business data (e.g., "2019-01-01") If NULL, no start date restriction is applied (data retrieved from the earliest available date). Default NULL.

end_period

character(1) end date of the data, in the same format as start_period. If NULL, no end date restriction is applied (data retrieved up to the most recent available date). Default NULL.

first_n

numeric(1) number of observations to retrieve from the start of the series. If NULL, no restriction is applied. Default NULL.

last_n

numeric(1) number of observations to retrieve from the end of the series. If NULL, no restriction is applied. Default NULL.

See also

Examples

# \donttest{
# fetch US dollar/Euro exchange rate
ecb_data("EXR", "D.USD.EUR.SP00.A")
#> Error in req_perform(req_error(req_url_query(req_url_path_append(req_user_agent(request("https://data-api.ecb.europa.eu/service/"),     "bbk (https://m-muecke.github.io/bbk)"), resource), ...),     body = ecb_error_body)): HTTP 404 Not Found.
#> • No results found.
#> • See docs at <https://data.ecb.europa.eu/help/api/status-codes>
# fetch data for multiple keys
ecb_data("EXR", c("D.USD", "JPY.EUR.SP00.A"))
#> Error in req_perform(req_error(req_url_query(req_url_path_append(req_user_agent(request("https://data-api.ecb.europa.eu/service/"),     "bbk (https://m-muecke.github.io/bbk)"), resource), ...),     body = ecb_error_body)): HTTP 500 Internal Server Error.
#> • <html lang="en"><head><title>European Central Bank</title><style type="text/css"> body{ font-family: "Droid Sans", sans-serif; margin:0; padding:0; background-color: #9099ab; text-align: center; } h1{  font-size:38px; font-weight: normal; color:#003299; }  @media screen and (max-width: 480px) { h1{ font-size: 30px; } }  p { font-size: 15px; line-height: 20px; color: #333; margin-bottom: 20px; }  p.signature{ font-style: italic; color:#DF0101; }  .centredFrame { margin: 0 auto; text-align: center; width: 70%;   } .content{ margin-top:40px; padding:40px; background-color: #fff; border:1px solid #CDE3E0; -moz-border-radius: 20px; border-radius: 20px; } </style><link href="//fonts.googleapis.com/css?family=Droid+Sans" rel="stylesheet" type="text/css"></head><body><div class="centredFrame"><div class="content"><img src="data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAASwAAACACAYAAACx1FRUAAAABGdBTUEAALGPC/xhBQAAACBjSFJNAAB6JQAAgIMAAPn/AACA6QAAdTAAAOpgAAA6mAAAF2+SX8VGAAAACXBIWXMAAC4jAAAuIwF4pT92AAAABmJLR0QA/wD/AP+gvaeTAAAfF0lEQVR42u2dd3wVRdfHN1QjEFroHZIQQkmBUFUQkd5LUBAsjwgIqICiCNKbNIEAoWkIvSNiA0VQRLGgovCIvSEIYnkEQRQ475zds8lkM7O7N0R4Sc4f309udsrOzu793TNnzswaRtxig2EyST3Bo4JkwTuCk4LTRuwSC/wct/i44C3BMsEIQTT3G5NpuBOYAKklmCr4VAAZqJcERo1kC/ysyhO3+KBgnKAa9yfDgsX8GzQQbNUIEBh1F1kUfRbqdJ9ogp9Tj6vLXRQ8I6jB/cuwYDFZQUHBXKOuEJiYJWDUelotPuJ4gQYLoN/wIfDBwSgT/IzHlGWwvtrLrDrrLr4gjk0Q5Ob+ZliwmMzSUvCNKTAoLLFLrL/RSzMKkDiWO2wF3D5yIAAYJvgZjynz11madhzrtY4fEsRzvzMsWEygPCBbQ/minoEKbZ6EJ+b0BqPmM5C33iJZaKxhX6U1ENl5Erzwyg0m+BmPOYeEefBvRAo8PqsPlL51JuSqvlzOc17Qh/ufYcFi/DI2VWDQohJWUqcBj8BHh8MBLl4PG3c0gcj2U8EIT0lvNdV+Giq0nm7+TfdZzlM9BSq0nAHJG1oB/F0QDhyIhmZ3jDbPYcSms8KG831gWLAY/2JlI6yrApHL4dNDNcyh3vyV7SGo1EZLzNLNEC4y85o+KwQ/11uUYeholNkASavamXV9eSQSCkUli7zJKt/Yg3w/GBYsRsdgpUNdWEkhDRbAHY8NhJ7D74eOQ4aKYVyK0yLyhxhG5hVlez86CO4d1Q8Gjv8PFG64QO/Mj1ucwPeFYcFinDRyE5o8ONsnrCyj9CZzdi9Xg4UZwxrQp4VWVdhKMCqsBaPcejCqrBbDQFGuzjJrZlDkzY0ChWJXcovpD7uu8fz0/rD04AxiBN8fhgWLsckjOGYO3zDUQIhIhtgpc4ZwqSU6MUsdQrXYcq4XtwQo9JZZ0LDPaGh+76NQpcNUyFdfiFv5dWCU2GyJnj1LiGXrpIY2pK8TLa7opXY7DvI9YliwGJs5plBEWn6k3CgSOOSLWeoW9AlGfJKVL3QLVGkxCyYu6g4ff1ITzp8pAQDBgnwA/xSBE99XgbXP3wS3DR9kOeFJ2JSR8CiMKJrRFPaAgmaljeD7xHAHMOGmSAjxCBFWTbeHHoTJC3pCQfE5uEliRse6LFYV15qze5MSe8D5P4ulxl+Z/F0E4GxxS7Sk40c+D4O+I/uLIeN6a9gYL4mWsNTy4tBTHB8x7U5odMcTYki5CoIwT6w5NCzK94sFi8nZbDRFSVhKCQ89AL+eKm2KzJxnOkHhRomWT0olVuXWQWEhdHvejksVo1M/VoTpSd2h2/3DIbLTZKjafirccvdIGDThHtj5amOAS4VS8+54rSGUakBDRVu0opIh9Ka58KSw1OCfgvDl11Wg0+BhEITDSEs4Z/L9YsFici4V5PioIo3mw08/VDIFpc/jA8DA0IUMi5sXmY70YDFU+1AM/ywByg8zlnSFYphWdKsYIm4Go/JqMKquMkMYjGJbTWssusMUSHm2ucif2yx39GRZaNh1kuWYt4eDZdfDsOl9zPS/fi8J5W5+ynLaW+f/WxDC940Fi8npMVc1kqFymyeh0wMPQav+D8Ot942w/FOKEAccsm1/5UZLrM6HQJdBw8AIec4cvqWuE1Qs3THKCvEquRm6DRoKF84VNcvf1G+EKVKp5cJWQktx/rtH9Yc7hWiWu3WmM0ZrAN83FiwmZ5Jui5ggdLBjSAIO09DxHZ+UMXShxCbo/dj9ZFnlFQI3FIxC2+UZPXDd0QHrLbINmt0zEpJWtbdmA9Mc65YFh7OOaJmFrYC8OMOYPuThTb5vLFhMzqOGcobOFgdVXFTkcsgrhOnY91VNwZqX0tEaAtZZqraqlKJFVFtpYYY0OIQOZydjpQXXGbekKcn3jwWLyVncH0h0uknxLdCLdmM4cawiBKMFJvuf/BAjfVbt4uCPrnz/WLCYnEWSpzCgIOGQrfRGK3Yq/0vw/M4bTMEaMLYfGMYey/+Ewze/YKR8xTVeG/t5MYnvHwsWk7PYlepX0m3MF5UM+RsshDFzb4cxiQkwKakbnPulDMDfBeCZDa3gkcl3wfj5CTA+0S89YeLizjB6zu2QF0Ma9OsHHXtnLXMe28T3jwWLyTkECb5OnfWzl8HIy2FoCDhyVq/0AaFnSwCcC01/LEDeey8GgtDKyihEGYei2B6MwMfI+LTF1vv5HrJgMdcuBcmJXp0WCquoSXFXmD+fEINfcUYuOGYpRHScAmOe6m2KQ3BDWoiMjvTqy2FuckfYe6AGvCSGgp98XAvgfGH49VhF2Lm7Ebz6egN47Q3/7H4zHnbvj4Ha3ScoN/bLMKMYvcx09A+ZeI+1YaD5YguzzIfStZema9Ndd6QZzW+tl+RnhQWL+X9AV5++n3mpghWz5JQRvgI6DhwOh46Em2v/Nm2/GcLaTUuLv8JhG4YhFH8WDGMv1O89xrSQnttxIxh5dlixV8W3+gODR9EPVnU1rSP08F9FpECtbhNg7fPN4NK5ovDqvjho0neUFXYRu+RHSYBW+LjuvySxZliwmKtMgk/BmkH58wtO4bKb/EIYPqGN+eYs7whGqU1p6wfR0hIWTs3Ok6F6y5kwbFpfgEu54Z13YyGq7TSo3XUiRPcY74+ECdCozyhrKxlc7uMVBiGEspAQ1OfIyf/RxzWhaA1Rrqbp9/pWEqy1Pq+9Ij8nLFjMtWVhLaP8eU3BqvU0FBJDwLtHDYA+jw2A7sOGpG3/gpTaCK0HPEy+J1xOcz0A7sRg7sYgPsN1tCuDHww4/1tpc62gEbHCn7NdCBZuGjgqMQH6jekHeVHorFCID9KthWQLiwWLyZaCtVEqcwQtqHy4mR6u1Su52XSCB9VfmLZfVaXVsHF7MyE2QQCnhUj9GQpwrrjpdP/n91Jw8Qwd8wPkgamLuqUtw/FBkBiS5sKhY+gWU6jyNU7dPeIN6Tr2sGCxYDHZU7D2SWVeNuSN+eIcG/OhMAixqNVtIkR1mQxVW86A6u2nwbdfhJvic9/Ye6FifBKEd5oM4R2nuIJO/agukyCXvTQnkJgrnEk0JwGWyVHvy6Xr+IwFiwWLyZ6C9RP5r7DMbM8lNPgX32ZTeY01q5fvZUhc0cEc3i3f0AaM4BesIFBc9OwF1oFhE/UyHSwqM1qaITzPgsWCxWRPwUJqU5k79fFPGgErtx7iej1hbicDl4KhTsJ4czG0cudQr0XQKF4RKZkVsFZ0DTf7zM+CxYLFXKOC1Z/KlFPHPi1VbytjbxFTbh28sruxaWV98GG0FaaAFlS8T9FCgULHvqBimyczriesS8NRFDP1iylOC66naxjHgsWCxWRvwdoilXsrTUiSLJHAMIauE63ti51DOPxcfh2EtZ8G8E+IKVqrtjW31ghifjdryR5i4pt0SmyGlVtawMbtN4NRclP6+jFGSwhmGL6oFduTcRi5Sdl+FiwWLCZbCtY5I23XzoGpx4VQRLSbCiNn94b979SFrg8+CGVaTbcExGmFhW6G/mP6py61Wf9CUyhQMxmMwtvMUIRUq8m2llB0cI+tkG1Qrdkc2LDD2vxvGwagimOpdYt8pW5+CroMfQDe2FcfWtz3CJRoPtv59ugu1PbygkssWCxYTPYWLKQvlStEAmZaM7XbToPj31c2xQSX3pTCkIcayRkFC2fshGU0fm7aOsMvvq4Kdz02EPJjOlpRRZ+1BAx3eqgurLYOU2BqYgL89UfJ1DJDp9xlBapK0e0YoHrwsBXIuv+9GKh2yyx5i+QTtA4S2/54ANfLgsWCxVzDgvWRVHZCanhDpTWwcGU7IT7l4NGZva3YLNUwz34lvRjaDRnbD+BSgVQROv5dFVi35VYYM/c2GD3zDkhM6QAfvBsLcDEkNc/JE2Wg18MPWHu+O9/KI4aXy9a2gd9/Kwwzl3W2BC8tKn6Y1O4fWLBYsJicIVhIC2nh9DkcxhW8cS6ECUsod9xSiLttrBU3pXsjcz2a6Su2Fep1nAJbdzYRwlXQdZeGY8cqwpSF3SC00QJrbSHGZMmCiIuvG82HmIRxUPSG+VC/9xOQu15qG36WluP0CPBaWbBYsJhrXLDkHQ8Gp+6LhT4onMXDRcZ+9mjHv/hyCTEMrN11AowQltna7c3g7f114f33Y2DHa43hyeSO0GPoAxCMi6kxah1nFc0N/BT1xdL2ydiOqqt0u4x+y4LFgsXkLMFCekl17HMPR0iyhAx9WqqXU6CjHWcKcUcGfMVXFO1hhTN96KfC19RjuEQs5Y+nWUndG6DT85zUzoczcZ0sWCxYTDYQLIxpKibNup3WWlFCiCq1ngGVW0+3hnNymmo5DQ4XcWcFnOGT/VQoTpgess0Mnyh8Q6L1/kOVxWXxvaAAtbGSYb2EggWLBYvJgYKFvCjVc4sqWBRf/9W6yyR4cU99eG5XQ2jVZaLl38rMSySERVX0prnQ9Z7H4I236sH8Na2gGcZ9YV0xyjfk1JXa904mr5EFiwWLySaChTwq1XVbemtJCFb4CuiHs4HkPP/PmH6QC0MNVIKFFhQOG3XrBqOegWLi75L1rai+66DL4GGWzyq9YF2QJgaQeZdxfSxYLFhMNhIspz+rZ7ohYaU1UKntk/DW23Vh7754qCw+Z9ji2HbaV14DRW6cC/lxm5qwFRnfBI2CFroFmgoL69AXVWHHriaQO4aWA6XlO0/Wnt2eBy/z2liwWLCYbCZYSG+pzua0u4NpLeE2Mfg5T3ySGfqQIQK+zjK4rkkiVBMsXt0OnpiXABUbJ1r7WTmHerWfhigxDAyKWwLlWs6A8rimMO0NOvhyjPpSO3pnwXWxYLFgMdlQsOQoeHuB9IupgaIoUlGEc2ZPpF0ftwimLO5KQ71guGt0f/XQEWcI8U04OJOI6WnxWGspLsw+/7AsuiYWLBYsJpsKFlD0u1z/XYb9WjC3mKySm6BFvxHm9slnTpaDUrfMtmKpvF9jf4iGoUYW+axYsFiwmBwkWMgLgjDpHHnJl/SVtkzYSrjp7pEQf9s4M1o9svOkjGsR03NYMMRxLbUEb2bxtbBgsWAx2VywkD8F9zvOlZusofWC/2UY6qFAYeAoop4pxAXMKYKOiusYmck4KxYsFiyGBSvdYum+ivOG0GzeI4IkwatG7OID4u97gvcJ3Ds+UTBc0MzAdyKmryMPDTmP/IvtZ8FiwWJykGDJvibcTz0iC9ocJRgv+PIKtJsFiwWLyYGCJbNbMI2CO6spLCeZYMN6XXx7w3r5xZtXuK0sWCxYTA4XLJlLNIuIWxbvFKwj8PN+Wg94NdvHgsWCxbBgXTOwYLFgMSxYLFgMCxbDgsWCxbBgsWCxYDEsWAwLFgsWw4LFsGCxYDEsWCxYLFgMCxZz1ejNouRJZX5OWLCY/x80o/V+uOf520w6cD3j64JS/JywYDEMw7BgMQzDsGAxDMOCxTAMw4LFMAzDgsUwDAsWwzAMCxbDMCxY3AkMw7BgMQzDsGAxDMOCxTAMw4LFMAzDgsUwDAsWwzAMCxbDMAwLFsMwOUawIgQ1BZEOogXlNAVx98ZYQZQgWJFejtKrC/Iq0oME1QS1pPNhG8IzcRExgh6CBEGczzLX0flqSOevbXjv+12Q8kVq+ivUx7mLUN46gmI+21uJrq2yJr0i9Xc49e3lPhhYRz1Bd+rXhtRnuvzB9CxEKfqljqHf7bMEtTtSkx5K5fFeFVCkF6f0SBfwGQtz6Ve5vH0NgfRVfqrDz/NjU5quu4ZHvwZCQeqnKMczXSbAegrT9URT//opU56up6om3Y8ehFEe5zMeQdeBz2AePPCb4KKh3gd7q6YB8yj9gqC+In0NpZ/TfMnwJn0muETY5/tHsI8a6dVJXQxr+1tnmw8KbvMoG0fXfMFR9g/BOo0II22ldqr6a7KPdu+X8h/w+UC8Qfn/FLRRpO+m9JNZ8AUYLvhccW3fCx7RlGlMfXlB0y9LNOUmS/05RpE+ltLxXt2iSB+nONc/ivuDz3huRfk90nMs5z9sWNtO++mvRKnczz4FYoF03npZJFht6Lt0QXHtzwjy+aznFansf32WeZbyn6UfOWf6Rko/rdED/L59Q3neko73EZyn4y+a99Bw37j/WU0Dl0p5GivSt0np1TSC9ZPLeX8ipdd10AjD+6UDo13K1/cou1FTrpNHuSkeN7aeokxLHw/Eh1L+cwqLwRbuvy9DsHIJXvK4vt80ZW/yKLdUU262I9/tjvSpUlorRfnxhr8XUPyuEawDLmUuulhmNtcrxHGsj75OlvI3zCLB6ujRBwt81FFLUa6Dj3KvOspEO9Jf8tAD7MdfHSJZSPC/DOVI9fDAKhIJJIQoqGngfKmiBor09dKvXRWNYH1FeT4gExmHB1ukenVWUjNH5+CvbF0aGo5yWIs3aeqoK+VZRuevSVaffVw1jGknpU+g4Z3cX8EeNzZJ8UA87+OB2Oco851jiGRbYL9ehmBNl+pHS+5REvYoGhZuFrxnmuUZyzZxfDGKSM9SIc1wzilINrEaQWqhGQaVltjoeC7t4yU1599LeU/QLz/e87k+f/RsC8DZ/qMk/n6fg/gsEqy2Up0TBUUFjaQv/c8+3AWzFNez28e5tzvKHKNnwE7fIv3YVtFYWD9Qnjfo2CapvjtlH5YtWPMC6JysFKxXpOOPS/Xe5TE8Qu5QpPeU0l/3IVgjpOPPS8cjPQRrQCZ8HfbDgxboSvp8yWXsrxMs54O07zIFq6RU7xkSf50vIr+HYI0L4LxTpT6wy/8i+U4mewiWkyVS/iI+8u+VfgDsY90DsJjtfv/EIXQ9rrJgDaJjaFWeomP7PMrnk0Y9OPxKkeqrHqBgAbk+nCMuP4K1mqxpu57XnE7309Iw7FUa179OY8nuV0CwUI1nCJ6WxqvnyZGncgie9TG+/kgy64t5CNYBOr/863xS88WUBQvH3Duor96g/nIz72+Xyrah67P/n+3zi3FY8rsgMx1+h8wKlty2xEyUb+KwMF6mfnmdHtyWHoKFQr5LquN9Sh8coGCtkPJXDUCwztAzMMsxDGnkUraalO8pEgf7+d17lQXrG/rxPUr/nzK8X3Mmv4y3u+NHbIFPwTpK99wut5zSVwYgWF+TxQt0rIhOsFTM1DRwoZSnnovT3Y9gqfwNMS6zOuc9/Ezy+UFz/roe4/1OmnrbeZS706VNuyXHvn3sXcmquN6HYL1L+T50WJkrLlOwhvi8Bj+CpeJhD8G6QDOdUxwTGFdKsFT+q64eZadI1mEthQO+zlUULCfnSIyDfIjOb4r++dPFPSSXPUbP5x7Hd2l2AIIl822G4bUkWHtpGCYT5UOwohXpa6X0cA/BOk438LiR9lrxGzTnDZHa+4VLBx6W6iriIViHyGJMprG/24MmC9YmRX+V15SrIpXbJ4WSyJbqvT4E66gUVvKnVPbEZQpWglRX0mUK1g5Fv1Tz4cOyh+AHpWPHr6CFtYiGdrYI9fWYoDgmWeN1SLQGOnyjV0uwNlO/D5K+C8hQTdkK0rD8PfrO1iSr0y472IdgXZB8i39Io6WjAQjWXsek3S6dYE0NoHMekyqcpDj5j1IDQzwEyzafazqGZHk0596lGKvL3Ov48nj5sAK57nY+JgWMTM5mfeBDsGRne2tFHZkVrLJSHX+5TOm3Iye2m2ANz4QPC2kuxSidUlzbvylY9g9BAcnlANQvgVozIE3xF/EhWFX/BcHqLR2vqf3yBzbr/pkPwTpPk2e6mWM/grWV/GmyxTVXJVhbHcFzboGKzulPfPCakhktTxW/6BK4aQvWOxrnXR9N2QaOc8+hczd1zHS5+SDqKjsjMMGaQ85Iub90D/g3jhgfvPlfkjXxh5TWxIdgyab5sCwSLIP8MCBZGNNJuJrSD4M9fI3wEKwURb9U8CFYTaXjN1xhwTou+SxnuPwYO+OO7C/yEbqfnzjCdR7yIVgJju9dvA9/k5dgzSMrKVqKmXRzo3zqeD4/p+v50OHTaxGAYBn03AQqWO9LsZKQQYQ9VPVdlw6a7OMXppaLYJ2QhmSqB/W/LlbWAB+/CIN8xmEtCeCh8IrDWqUo01lKX+8YVhg0o2Snv+ARh/WLwpew3BE0mVnByq2Ip1FRNhNxWDt9xGE1d6T1c9TR2sc1yBMnYT7yH5ACGgsrrM3TinhAeXj/uRS5bd/P6o4hbT6POCwVmZn46Ojj3jVXlGsppT+neD7ba2fs1HFYJVzcR7o4rGDynTn14G5H2RsM+kX4iwBH9Pkej056SBr3yyq63cVxbgvWJ5pz7JXO7xZt3JI6Sg7cu0Bmr9fDHUvXe8llYkFFazrfGcn5L/eXym+RIjkum2mWJRykOk6T81k3DP5U45zfJc2qXG6k+whpltUZRnGn5kekId33s5rnaJvHUPmcxm85i4Jhsa9v9tH2pVJfV/GRfyfl/4rixQzHrBbyH0eZB6UfB93wd5UUD6gKvJwj+c4uKPprZibuW0vqp7OOZ/Mc+aX6eFh75zSrKOTA5PMay3WjZOEXd7HATrlEun+m0YM59AxYq1jIUVyFzHaZMJeAO+cXLpoEKsZlGOksU4ZM4VKKxkfQL5WftXlVpXNX83lz81E7IyjAzu9DEUznqKDpL1UIhX2eii71FpKuubBm7WYk9VmQxgkcRXFSWbGW0KD1W3a/eq3xzE8zuBU1/RLqsm6tOvVRfpfwgao+hbiY1Nd5fMafRZJVJc9G5aF6whXractSWjWXvs5H1x2u8fkVpetWPUfVfcaQqYyAKnQf7Loq0v9u5SrQ9bjlK0B5IjRtC6V2l3cJmo2g9DwaPSir0QP7O17VXHvJK8AZhrlm4E5gGIYFi2EYhgWLYRgWLIZhGBYshmEYFiyGYViwGIZhWLAYhmFYsBiGYcFiGIZhwWIYhmHByp7g9iW47QzubvEh86+Au2DM4GeNBYu5fHB1O24Bglt4bGL+FbBvh/OzxoLFMAzDgsUwDAsWwzAMCxbDMAwLFsMwLFgMwzAsWAzDMCxYDMOwYDEMw7BgMQzDsGAxDMOCxTAMw4KV3Wks6CzoRODnbob12nln3vqC9oa1Q4MzraCgq2G9kl51nkaCLoJegjZG4K87jxUMEYwS3GVYrzD3Uw5fIf6QYKQgXrN4u7eglaZ8jGCQYb1mXZXeTDBGMEDTZ4XpujsTHQRtBR2lvm6qKFeC7kcH6f50o/Op2tFGU4/cf3h/rudnngXrWuaIABT0UeR9h9JCFWk1KW2Z5jyfOer/h4THj+Ds1rRxrkfZ7ooyIx15cglWUNp4R1qYVE4lkCsV9Vdy5GmgabvML4q622vynlbkDRL8TukNFOl4HRcpvRo/8yxY1zI/CP4W9BQkCG4T9BWUU+Td4yJYNShtoeY8JwU/kpWGYvgb5Y90aVsp+jJjvnVkmcSQRfQBHd+kKYsW3CXB54KbycJ7UbBe84X/lOprLAnZd3SsnqJMB0pLFtQl6+cQtdHZjl7UryigG6jcBPof+/pWRf3tKN8KyteT8rbStP8Q5T9D1q6c/j2l/UU/APzcs2Bds/xE+Mn7Cj34xRVpkZQ2T1P2V8F7jiEM5nfb52kX5blbk77dJd22+B5yHC+jqasM5f9DUEAwm/4frMn/FKXncxwv5tGH/alcuEe+tpTvdh/3BQXrK+lH4DUpbSMd+5l+mKrwM8+Cda0L1okrIFj4ZXpf+n8g5R+gyW8Px/a6tCcfWVFHFWmhNAzCa4vzeX2d6Jxf0N9tLnl7Up61guAA+nsklWvqU7Du8ClY5wVryJoD8rslSJ9n0mcWLBasa96HhQ/7JMFEwVTBFEH5LBasz8nKeoyslzM0HC2tyX831efl53qd8hVSpPWRfD/jaJjn1R/TKf83mskF2fe1ifIeo2FbVgpWa8q3QzBWMJna1lcjWGg9veCwTJH9dGwRCxYLVnbgXY1zt3EWC9Z+xTnc/FcPUp4OHu1f49Img2bY7KHS2xphU13HZJ/9N0G6nvFZKFitNPdlr4uFtU+ambTz20PgzSxYLFjZge/oCx1BfpUa9KW9LkBrpgqlLdKc5xjNFNYSJFLe6S7t6kh5Rnu03559vM4lTxFpRm+NR332rN7UAPqwoeAUlWuTxUPCiTQ8xvtTRzMZYgvWXsekQA/p/w0sWCxYOc3pbv9Kh7g4uSe7ON1lH9ZzlL+9Jn+INNxyiyFzOplt8ruEZeT1EB8vMbVFQuW0fz6LBatvAD4sN38fCxYLVrYRrJM+895OD/3TCue3HSvVUFMWZ6kOOgJNL9Bx3flGUZ0vKwJNMQj0HKXHaSylBYLc0rGXKX+eLBCsOwX9HMeAhs1ZKVi9WLBYsJj08VH4IK8SpAhW07CphSb/a5JTGgVhqeB/dGy5y3lAMZtnO9ZTXMqtlvwxGEM1R4oHcwt5iKD0E+Rn2kr/r/TojxspX5JHvqGSQE2QYrZae5SbRPlu9Tlj+V/qY/u+LFIIbhDlPexS307KE8bPPAvWtcw6Ciy8SFzwCDdAi+VJhyP4Z3KSu53nLcGziuM4NDzrsqTHoJnCrxznxCDQJh5WxwDJ4W4LXl6PdkaTsD7ska+EFCFvM9RHf99Hw9x4j3xN6EfhFwrdsO/LSUXsF17rx/SjY7jEjX2nmf1lsoj/AzSVcRgd4vznAAAAJXRFWHRkYXRlOmNyZWF0ZQAyMDE4LTA2LTA1VDEyOjI1OjQxKzAyOjAw751EtwAAACV0RVh0ZGF0ZTptb2RpZnkAMjAxOC0wNi0wNVQxMjoyNDo1NCswMjowMO+QuAwAAAAASUVORK5CYII="><h1>ECB Data <br> <br> ECB Data Portal gives an error, please try with our home page https://data.ecb.europa.eu <br> <br> If our home page brought you here, we are working to make it again available for you. Please try again after few minutes.</h1>We apologize for any inconvenience this might have caused.<p class="signature">HTTP status code: 500</p></div></div></body></html>
#> • See docs at <https://data.ecb.europa.eu/help/api/status-codes>
# }