Fetch European Central Bank (ECB) dimensions

Retrieve the dimension structure for a given dataflow from the ECB SDMX Web Service.

Usage

ecb_dimension(id)

Source

https://data.ecb.europa.eu/help/api/metadata

Arguments

id

(character(1))
The id of the data structure definition to query (e.g., "ECB_EXR1").

Value

A data.table::data.table() with columns:

id

The dimension id (e.g., "FREQ", "CURRENCY")

position

The position of the dimension in the series key

codelist

The id of the associated codelist (e.g., "CL_FREQ")

See also

Examples

# \donttest{
ecb_dimension("ECB_EXR1")
#>                id position      codelist
#>            <char>    <int>        <char>
#> 1:           FREQ        1       CL_FREQ
#> 2:       CURRENCY        2   CL_CURRENCY
#> 3: CURRENCY_DENOM        3   CL_CURRENCY
#> 4:       EXR_TYPE        4   CL_EXR_TYPE
#> 5:     EXR_SUFFIX        5 CL_EXR_SUFFIX
# }