Euro foreign exchange reference rates

Fetch the latest or historical Euro foreign exchange reference rates from the European Central Bank.

Usage

ecb_euro_rates(x = c("latest", "history"))

Source

https://www.ecb.europa.eu/stats/policy_and_exchange_rates/euro_reference_exchange_rates/html/index.en.html

Arguments

x

(character(1)) one of "latest" or "history". Default "latest".

Value

A data.table::data.table() with the reference rates.

Details

Note you can achieve the same by calling the ecb_data() function with the right parameters for each currency.

The reference rates are usually updated at around 16:00 CET every working day, except on TARGET closing days.

They are based on the daily concertation procedure between central banks across Europe, which normally takes place around 14:10 CET. The reference rates are published for information purposes only. Using the rates for transaction purposes is strongly discouraged.

Examples

# \donttest{
ecb_euro_rates()
#>           date currency       rate
#>         <Date>   <char>      <num>
#>  1: 2025-06-13      USD     1.1512
#>  2: 2025-06-13      JPY   165.9400
#>  3: 2025-06-13      BGN     1.9558
#>  4: 2025-06-13      CZK    24.8310
#>  5: 2025-06-13      DKK     7.4586
#>  6: 2025-06-13      GBP     0.8505
#>  7: 2025-06-13      HUF   403.4300
#>  8: 2025-06-13      PLN     4.2745
#>  9: 2025-06-13      RON     5.0330
#> 10: 2025-06-13      SEK    10.9635
#> 11: 2025-06-13      CHF     0.9359
#> 12: 2025-06-13      ISK   144.2000
#> 13: 2025-06-13      NOK    11.4435
#> 14: 2025-06-13      TRY    45.3974
#> 15: 2025-06-13      AUD     1.7777
#> 16: 2025-06-13      BRL     6.4007
#> 17: 2025-06-13      CAD     1.5690
#> 18: 2025-06-13      CNY     8.2682
#> 19: 2025-06-13      HKD     9.0362
#> 20: 2025-06-13      IDR 18783.6700
#> 21: 2025-06-13      ILS     4.1422
#> 22: 2025-06-13      INR    99.1100
#> 23: 2025-06-13      KRW  1577.4200
#> 24: 2025-06-13      MXN    21.8807
#> 25: 2025-06-13      MYR     4.8874
#> 26: 2025-06-13      NZD     1.9167
#> 27: 2025-06-13      PHP    64.7060
#> 28: 2025-06-13      SGD     1.4777
#> 29: 2025-06-13      THB    37.3510
#> 30: 2025-06-13      ZAR    20.7360
#>           date currency       rate
# }