Returns the available ECB metadata

Retrieval of the metadata stored in the ECB's time series database. Access via the SDMX Web Service API of the ECB

Usage

ecb_metadata(type, agency = NULL, id = NULL)

Source

https://data.ecb.europa.eu/help/api/metadata

Arguments

type

(character(1)) the type of metadata to query. One of: "datastructure", "dataflow", "codelist", or "concept".

agency

(character(1)) id of the agency to query. Default NULL.

id

(character(1)) id of the resource to query. Default NULL.

Value

A data.table::data.table() with the queried metadata. The columns are:

agency

The agency of the metadata

id

The id of the metadata

name

The name of the metadata

See also

Examples

# \donttest{
ecb_metadata("datastructure")
#>       agency              id
#>       <char>          <char>
#>  1:      ECB        ECB_AME1
#>  2:      ECB        ECB_BCS1
#>  3:      ECB        ECB_BKN1
#>  4:      ECB        ECB_BLS1
#>  5:      ECB        ECB_BOP1
#>  6:      ECB     ECB_BOP_BNT
#>  7:      ECB        ECB_BSI1
#>  8:      ECB        ECB_CAR1
#>  9:      ECB        ECB_CBD1
#> 10:      ECB        ECB_CBD2
#> 11:      ECB        ECB_CCP1
#> 12:      ECB        ECB_CPP3
#> 13:      ECB        ECB_DCM1
#> 14:      ECB         ECB_DD1
#> 15:      ECB        ECB_DWA1
#> 16:      ECB        ECB_EON1
#> 17:      ECB        ECB_ESA1
#> 18:      ECB        ECB_EST1
#> 19:      ECB        ECB_EWT1
#> 20:      ECB        ECB_EXR1
#> 21:      ECB        ECB_FCT1
#> 22:      ECB        ECB_FMD2
#> 23:      ECB        ECB_FVC1
#> 24:      ECB        ECB_FXS1
#> 25:      ECB        ECB_GST1
#> 26:      ECB        ECB_ICO1
#> 27:      ECB        ECB_ICP1
#> 28:      ECB        ECB_ICP2
#> 29:      ECB       ECB_ICPF1
#> 30:      ECB        ECB_IFI1
#> 31:      ECB        ECB_ILM1
#> 32:      ECB        ECB_IRS1
#> 33:      ECB        ECB_IVF1
#> 34:      ECB        ECB_LIG1
#> 35:      ECB        ECB_MFI1
#> 36:      ECB        ECB_MIR1
#> 37:      ECB        ECB_MMS1
#> 38:      ECB       ECB_MMSR1
#> 39:      ECB        ECB_MPD1
#> 40:      ECB        ECB_OFI1
#> 41:      ECB        ECB_OMO1
#> 42:      ECB        ECB_PAY1
#> 43:      ECB       ECB_PAY10
#> 44:      ECB       ECB_PAY11
#> 45:      ECB       ECB_PAY12
#> 46:      ECB       ECB_PAY13
#> 47:      ECB       ECB_PAY14
#> 48:      ECB        ECB_PAY2
#> 49:      ECB        ECB_PAY3
#> 50:      ECB        ECB_PAY4
#> 51:      ECB        ECB_PAY5
#> 52:      ECB        ECB_PAY6
#> 53:      ECB        ECB_PAY7
#> 54:      ECB        ECB_PAY8
#> 55:      ECB        ECB_PAY9
#> 56:      ECB        ECB_PSS1
#> 57:      ECB        ECB_RAI1
#> 58:      ECB        ECB_RES1
#> 59:      ECB        ECB_RIR2
#> 60:      ECB        ECB_RPP1
#> 61:      ECB        ECB_RTD1
#> 62:      ECB        ECB_SAFE
#> 63:      ECB        ECB_SEC1
#> 64:      ECB        ECB_SEE1
#> 65:      ECB        ECB_SHI1
#> 66:      ECB        ECB_SHS6
#> 67:      ECB        ECB_SSI1
#> 68:      ECB        ECB_SSS1
#> 69:      ECB        ECB_STP1
#> 70:      ECB        ECB_STS1
#> 71:      ECB        ECB_SUP1
#> 72:      ECB        ECB_SUR1
#> 73:      ECB        ECB_SUR2
#> 74:      ECB        ECB_TGB1
#> 75:      ECB        ECB_TRD1
#> 76:      ECB        ECB_WTS1
#> 77:      ECB EUROSTAT_BOP_01
#> 78:    ESTAT             JVS
#> 79:    ESTAT             LCI
#> 80:    ESTAT         NA_MAIN
#> 81:    ESTAT          NA_SEC
#> 82: EUROSTAT    ESTAT_ESAIEA
#> 83: EUROSTAT   EUROSTAT_JVC2
#> 84: EUROSTAT   EUROSTAT_LFS1
#> 85:      IMF             BOP
#> 86:      IMF         BOP1_15
#>       agency              id
#>                                                                      name
#>                                                                    <char>
#>  1:                                                                 AMECO
#>  2:                                             Business Cycle Statistics
#>  3:                                                  Banknotes Statistics
#>  4:                                        Bank Lending Survey Statistics
#>  5:             Balance of Payments, I.I.P. and Reserve Assets Statistics
#>  6:                             Statistics on Shipments of Euro Banknotes
#>  7:                                                   Balance Sheet Items
#>  8:                                                        Car Statistics
#>  9:                               Statistics on Consolidated Banking Data
#> 10:                               Statistics on Consolidated Banking Data
#> 11:                              Central Counterparty Clearing Statistics
#> 12:                                  Commercial property price statistics
#> 13:                                                              Dealogic
#> 14:                                                          Derived Data
#> 15:                                        Distributional Wealth Accounts
#> 16:                                                   Internal Eonia Rate
#> 17:                                               ESA95 National Accounts
#> 18:                                                  Euro Short-Term Rate
#> 19:                                                      ECB Wage Tracker
#> 20:                                                        Exchange Rates
#> 21:                                                              Forecast
#> 22:               Financial market data (not related to foreign exchange)
#> 23:                  Financial Vehicle Corporation Assets and Liabilities
#> 24:  Foreign eXchange Statistics (foreign reserves & exchange operations)
#> 25:                                                 Government Statistics
#> 26:                                     Insurance corporations operations
#> 27:                                            Indices of Consumer Prices
#> 28:                                            Indices of Consumer Prices
#> 29:       Insurance corporations and pension funds assets and liabilities
#> 30:                                   Indicators of Financial Integration
#> 31:                                         Internal Liquidity Management
#> 32:              Interest Rate Statistics (2004 EU Member States & ACCBs)
#> 33:                                                      Investment Funds
#> 34:                                      Large Insurance Group Statistics
#> 35:                         MFI - List of monetary financial institutions
#> 36:                                          MFI Interest Rate Statistics
#> 37:                                                   Money Market Survey
#> 38:                                               Money Market Statistics
#> 39:                                     Macroeconomic Projection database
#> 40:                                        Other Financial Intermediaries
#> 41:                                                Open market operations
#> 42:                                Payments transactions (Key indicators)
#> 43:                    Electronic card payments sent by merchant category
#> 44:   Number of cards issued by resident PSPs by card function and scheme
#> 45:    Number of terminals provided by resident PSPs by type and function
#> 46:                            Number of participants in payments systems
#> 47:                                      Transactions in payments systems
#> 48:                               Credit transfers (including fraud data)
#> 49:                                  Direct debits (including fraud data)
#> 50:                   E-money payment transactions (including fraud data)
#> 51: Card payments and cash withdrawals using cards (including fraud data)
#> 52:                    Payment initiation services (including fraud data)
#> 53:                               Losses due to fraud by liability bearer
#> 54:                                ATM, OTC and POS terminal transactions
#> 55:                                 Payments transactions by PSP category
#> 56:                                        Payment and Settlement Systems
#> 57:                                            Risk Assessment Indicators
#> 58:                                                Real Estate Statistics
#> 59:                                                 Retail Interest Rates
#> 60:                                 Residential Property Price Indicators
#> 61:      Real Time Database (context of Euro Area Business Cycle Network)
#> 62:       Survey on the Access to Finance of Small and Medium Enterprises
#> 63:                                                            Securities
#> 64:                              Securities Exchange (trading) Statistics
#> 65:                              Structural Housing Indicators Statistics
#> 66:                                         Securities Holding Statistics
#> 67:                             Banking Structural Statistical Indicators
#> 68:                                      Securities Settlement Statistics
#> 69:                           Short-Term European Paper Statistics (STEP)
#> 70:                                                 Short-Term Statistics
#> 71:                                        Supervisory Banking Statistics
#> 72:                                                       Opinion Surveys
#> 73:                                                       Opinion Surveys
#> 74:                                                       Target Balances
#> 75:                                                        External Trade
#> 76:          Overall, Import, Export and Double Export Weights Statistics
#> 77:                                          Eurostat Balance of Payments
#> 78:                                                Job Vacancy Statistics
#> 79:                                                     Labour Cost Index
#> 80:                                      National Account Main Aggregates
#> 81:                                              National Sector Accounts
#> 82:                                                                Sector
#> 83:                                                Job Vacancy Statistics
#> 84:                                                   Labour Force Survey
#> 85:                        Balance of Payments, I.I.P. and Reserve Assets
#> 86:                 Balance of Payments, I.I.P. and Reserve Assets, v1.15
#>                                                                      name
ecb_metadata("datastructure", "ECB")
#>     agency              id
#>     <char>          <char>
#>  1:    ECB        ECB_AME1
#>  2:    ECB        ECB_BCS1
#>  3:    ECB        ECB_BKN1
#>  4:    ECB        ECB_BLS1
#>  5:    ECB        ECB_BOP1
#>  6:    ECB     ECB_BOP_BNT
#>  7:    ECB        ECB_BSI1
#>  8:    ECB        ECB_CAR1
#>  9:    ECB        ECB_CBD1
#> 10:    ECB        ECB_CBD2
#> 11:    ECB        ECB_CCP1
#> 12:    ECB        ECB_CPP3
#> 13:    ECB        ECB_DCM1
#> 14:    ECB         ECB_DD1
#> 15:    ECB        ECB_DWA1
#> 16:    ECB        ECB_EON1
#> 17:    ECB        ECB_ESA1
#> 18:    ECB        ECB_EST1
#> 19:    ECB        ECB_EWT1
#> 20:    ECB        ECB_EXR1
#> 21:    ECB        ECB_FCT1
#> 22:    ECB        ECB_FMD2
#> 23:    ECB        ECB_FVC1
#> 24:    ECB        ECB_FXS1
#> 25:    ECB        ECB_GST1
#> 26:    ECB        ECB_ICO1
#> 27:    ECB        ECB_ICP1
#> 28:    ECB        ECB_ICP2
#> 29:    ECB       ECB_ICPF1
#> 30:    ECB        ECB_IFI1
#> 31:    ECB        ECB_ILM1
#> 32:    ECB        ECB_IRS1
#> 33:    ECB        ECB_IVF1
#> 34:    ECB        ECB_LIG1
#> 35:    ECB        ECB_MFI1
#> 36:    ECB        ECB_MIR1
#> 37:    ECB        ECB_MMS1
#> 38:    ECB       ECB_MMSR1
#> 39:    ECB        ECB_MPD1
#> 40:    ECB        ECB_OFI1
#> 41:    ECB        ECB_OMO1
#> 42:    ECB        ECB_PAY1
#> 43:    ECB       ECB_PAY10
#> 44:    ECB       ECB_PAY11
#> 45:    ECB       ECB_PAY12
#> 46:    ECB       ECB_PAY13
#> 47:    ECB       ECB_PAY14
#> 48:    ECB        ECB_PAY2
#> 49:    ECB        ECB_PAY3
#> 50:    ECB        ECB_PAY4
#> 51:    ECB        ECB_PAY5
#> 52:    ECB        ECB_PAY6
#> 53:    ECB        ECB_PAY7
#> 54:    ECB        ECB_PAY8
#> 55:    ECB        ECB_PAY9
#> 56:    ECB        ECB_PSS1
#> 57:    ECB        ECB_RAI1
#> 58:    ECB        ECB_RES1
#> 59:    ECB        ECB_RIR2
#> 60:    ECB        ECB_RPP1
#> 61:    ECB        ECB_RTD1
#> 62:    ECB        ECB_SAFE
#> 63:    ECB        ECB_SEC1
#> 64:    ECB        ECB_SEE1
#> 65:    ECB        ECB_SHI1
#> 66:    ECB        ECB_SHS6
#> 67:    ECB        ECB_SSI1
#> 68:    ECB        ECB_SSS1
#> 69:    ECB        ECB_STP1
#> 70:    ECB        ECB_STS1
#> 71:    ECB        ECB_SUP1
#> 72:    ECB        ECB_SUR1
#> 73:    ECB        ECB_SUR2
#> 74:    ECB        ECB_TGB1
#> 75:    ECB        ECB_TRD1
#> 76:    ECB        ECB_WTS1
#> 77:    ECB EUROSTAT_BOP_01
#>     agency              id
#>                                                                      name
#>                                                                    <char>
#>  1:                                                                 AMECO
#>  2:                                             Business Cycle Statistics
#>  3:                                                  Banknotes Statistics
#>  4:                                        Bank Lending Survey Statistics
#>  5:             Balance of Payments, I.I.P. and Reserve Assets Statistics
#>  6:                             Statistics on Shipments of Euro Banknotes
#>  7:                                                   Balance Sheet Items
#>  8:                                                        Car Statistics
#>  9:                               Statistics on Consolidated Banking Data
#> 10:                               Statistics on Consolidated Banking Data
#> 11:                              Central Counterparty Clearing Statistics
#> 12:                                  Commercial property price statistics
#> 13:                                                              Dealogic
#> 14:                                                          Derived Data
#> 15:                                        Distributional Wealth Accounts
#> 16:                                                   Internal Eonia Rate
#> 17:                                               ESA95 National Accounts
#> 18:                                                  Euro Short-Term Rate
#> 19:                                                      ECB Wage Tracker
#> 20:                                                        Exchange Rates
#> 21:                                                              Forecast
#> 22:               Financial market data (not related to foreign exchange)
#> 23:                  Financial Vehicle Corporation Assets and Liabilities
#> 24:  Foreign eXchange Statistics (foreign reserves & exchange operations)
#> 25:                                                 Government Statistics
#> 26:                                     Insurance corporations operations
#> 27:                                            Indices of Consumer Prices
#> 28:                                            Indices of Consumer Prices
#> 29:       Insurance corporations and pension funds assets and liabilities
#> 30:                                   Indicators of Financial Integration
#> 31:                                         Internal Liquidity Management
#> 32:              Interest Rate Statistics (2004 EU Member States & ACCBs)
#> 33:                                                      Investment Funds
#> 34:                                      Large Insurance Group Statistics
#> 35:                         MFI - List of monetary financial institutions
#> 36:                                          MFI Interest Rate Statistics
#> 37:                                                   Money Market Survey
#> 38:                                               Money Market Statistics
#> 39:                                     Macroeconomic Projection database
#> 40:                                        Other Financial Intermediaries
#> 41:                                                Open market operations
#> 42:                                Payments transactions (Key indicators)
#> 43:                    Electronic card payments sent by merchant category
#> 44:   Number of cards issued by resident PSPs by card function and scheme
#> 45:    Number of terminals provided by resident PSPs by type and function
#> 46:                            Number of participants in payments systems
#> 47:                                      Transactions in payments systems
#> 48:                               Credit transfers (including fraud data)
#> 49:                                  Direct debits (including fraud data)
#> 50:                   E-money payment transactions (including fraud data)
#> 51: Card payments and cash withdrawals using cards (including fraud data)
#> 52:                    Payment initiation services (including fraud data)
#> 53:                               Losses due to fraud by liability bearer
#> 54:                                ATM, OTC and POS terminal transactions
#> 55:                                 Payments transactions by PSP category
#> 56:                                        Payment and Settlement Systems
#> 57:                                            Risk Assessment Indicators
#> 58:                                                Real Estate Statistics
#> 59:                                                 Retail Interest Rates
#> 60:                                 Residential Property Price Indicators
#> 61:      Real Time Database (context of Euro Area Business Cycle Network)
#> 62:       Survey on the Access to Finance of Small and Medium Enterprises
#> 63:                                                            Securities
#> 64:                              Securities Exchange (trading) Statistics
#> 65:                              Structural Housing Indicators Statistics
#> 66:                                         Securities Holding Statistics
#> 67:                             Banking Structural Statistical Indicators
#> 68:                                      Securities Settlement Statistics
#> 69:                           Short-Term European Paper Statistics (STEP)
#> 70:                                                 Short-Term Statistics
#> 71:                                        Supervisory Banking Statistics
#> 72:                                                       Opinion Surveys
#> 73:                                                       Opinion Surveys
#> 74:                                                       Target Balances
#> 75:                                                        External Trade
#> 76:          Overall, Import, Export and Double Export Weights Statistics
#> 77:                                          Eurostat Balance of Payments
#>                                                                      name
ecb_metadata("datastructure", "ECB", "ECB_EXR1")
#>    agency       id           name
#>    <char>   <char>         <char>
#> 1:    ECB ECB_EXR1 Exchange Rates
ecb_metadata("datastructure", id = "ECB_EXR1")
#>    agency       id           name
#>    <char>   <char>         <char>
#> 1:    ECB ECB_EXR1 Exchange Rates
# }