Fetch Norges Bank (NoB) data

Retrieve time series data from the Norges Bank SDMX Web Service.

Usage

nob_data(
  flow,
  key = NULL,
  start_period = NULL,
  end_period = NULL,
  first_n = NULL,
  last_n = NULL
)

Source

https://www.norges-bank.no/en/topics/Statistics/open-data/

Arguments

flow

(character(1))
The dataflow to query. See nob_metadata() for available dataflows.

key

(NULL | character(1))
The series key to query using dot-separated dimension values (e.g., "B.USD.NOK.SP"). Use + for multiple values in one dimension (e.g., "B.USD+EUR.NOK.SP"). If NULL, all data for the flow is returned. Default NULL.

start_period

(NULL | character(1) | integer(1))
Start date of the data (e.g., "2024-01-01" or 2024). If NULL, no start date restriction is applied. Default NULL.

end_period

(NULL | character(1) | integer(1))
End date of the data, in the same format as start_period. If NULL, no end date restriction is applied. Default NULL.

first_n

(NULL | numeric(1))
Number of observations to retrieve from the start of the series. If NULL, no restriction is applied. Default NULL.

last_n

(NULL | numeric(1))
Number of observations to retrieve from the end of the series. If NULL, no restriction is applied. Default NULL.

Value

A data.table::data.table() with the requested data.

See also

Examples

# \donttest{
# fetch USD/NOK exchange rate
nob_data("EXR", "B.USD.NOK.SP", last_n = 5L)
#>          date          key  value   freq base_cur quote_cur  tenor
#>        <Date>       <char>  <num> <char>   <char>    <char> <char>
#> 1: 2026-03-10 B.USD.NOK.SP 9.5756  daily      USD       NOK     SP
#> 2: 2026-03-11 B.USD.NOK.SP 9.6451  daily      USD       NOK     SP
#> 3: 2026-03-12 B.USD.NOK.SP 9.6627  daily      USD       NOK     SP
#> 4: 2026-03-13 B.USD.NOK.SP 9.7238  daily      USD       NOK     SP
#> 5: 2026-03-16 B.USD.NOK.SP 9.7134  daily      USD       NOK     SP

# fetch multiple exchange rates
nob_data("EXR", "B.USD+EUR+GBP.NOK.SP", start_period = "2024-01-01")
#>             date          key   value   freq base_cur quote_cur  tenor
#>           <Date>       <char>   <num> <char>   <char>    <char> <char>
#>    1: 2024-01-02 B.USD.NOK.SP 10.2971  daily      USD       NOK     SP
#>    2: 2024-01-03 B.USD.NOK.SP 10.3672  daily      USD       NOK     SP
#>    3: 2024-01-04 B.USD.NOK.SP 10.3027  daily      USD       NOK     SP
#>    4: 2024-01-05 B.USD.NOK.SP 10.3553  daily      USD       NOK     SP
#>    5: 2024-01-08 B.USD.NOK.SP 10.3800  daily      USD       NOK     SP
#>   ---                                                                 
#> 1658: 2026-03-10 B.EUR.NOK.SP 11.1470  daily      EUR       NOK     SP
#> 1659: 2026-03-11 B.EUR.NOK.SP 11.1700  daily      EUR       NOK     SP
#> 1660: 2026-03-12 B.EUR.NOK.SP 11.1575  daily      EUR       NOK     SP
#> 1661: 2026-03-13 B.EUR.NOK.SP 11.1590  daily      EUR       NOK     SP
#> 1662: 2026-03-16 B.EUR.NOK.SP 11.1490  daily      EUR       NOK     SP

# fetch policy rate
nob_data("IR", last_n = 5L)
#>           date         key value    freq instrument_type  tenor unit_measure
#>         <Date>      <char> <num>  <char>          <char> <char>       <char>
#>  1: 1975-07-15 A.KPRA.SD.R  0.08  annual            KPRA     SD            R
#>  2: 1975-07-16 A.KPRA.SD.R  1.33  annual            KPRA     SD            R
#>  3: 1975-07-17 A.KPRA.SD.R  3.54  annual            KPRA     SD            R
#>  4: 1975-07-18 A.KPRA.SD.R  4.50  annual            KPRA     SD            R
#>  5: 1975-07-19 A.KPRA.SD.R  4.29  annual            KPRA     SD            R
#>  6: 2026-03-10 B.KPRA.OL.R  5.00   daily            KPRA     OL            R
#>  7: 2026-03-11 B.KPRA.OL.R  5.00   daily            KPRA     OL            R
#>  8: 2026-03-12 B.KPRA.OL.R  5.00   daily            KPRA     OL            R
#>  9: 2026-03-13 B.KPRA.OL.R  5.00   daily            KPRA     OL            R
#> 10: 2026-03-16 B.KPRA.OL.R  5.00   daily            KPRA     OL            R
#> 11: 2026-03-10 B.KPRA.SD.R  4.00   daily            KPRA     SD            R
#> 12: 2026-03-11 B.KPRA.SD.R  4.00   daily            KPRA     SD            R
#> 13: 2026-03-12 B.KPRA.SD.R  4.00   daily            KPRA     SD            R
#> 14: 2026-03-13 B.KPRA.SD.R  4.00   daily            KPRA     SD            R
#> 15: 2026-03-16 B.KPRA.SD.R  4.00   daily            KPRA     SD            R
#> 16: 2026-03-10 B.KPRA.RR.R  3.00   daily            KPRA     RR            R
#> 17: 2026-03-11 B.KPRA.RR.R  3.00   daily            KPRA     RR            R
#> 18: 2026-03-12 B.KPRA.RR.R  3.00   daily            KPRA     RR            R
#> 19: 2026-03-13 B.KPRA.RR.R  3.00   daily            KPRA     RR            R
#> 20: 2026-03-16 B.KPRA.RR.R  3.00   daily            KPRA     RR            R
#> 21: 2025-10-01 M.KPRA.OL.R  5.00 monthly            KPRA     OL            R
#> 22: 2025-11-01 M.KPRA.OL.R  5.00 monthly            KPRA     OL            R
#> 23: 2025-12-01 M.KPRA.OL.R  5.00 monthly            KPRA     OL            R
#> 24: 2026-01-01 M.KPRA.OL.R  5.00 monthly            KPRA     OL            R
#> 25: 2026-02-01 M.KPRA.OL.R  5.00 monthly            KPRA     OL            R
#> 26: 1975-07-15 A.KPRA.OL.R  1.08  annual            KPRA     OL            R
#> 27: 1975-07-16 A.KPRA.OL.R  2.33  annual            KPRA     OL            R
#> 28: 1975-07-17 A.KPRA.OL.R  4.54  annual            KPRA     OL            R
#> 29: 1975-07-18 A.KPRA.OL.R  5.50  annual            KPRA     OL            R
#> 30: 1975-07-19 A.KPRA.OL.R  5.29  annual            KPRA     OL            R
#> 31: 2025-10-01 M.KPRA.RR.R  3.00 monthly            KPRA     RR            R
#> 32: 2025-11-01 M.KPRA.RR.R  3.00 monthly            KPRA     RR            R
#> 33: 2025-12-01 M.KPRA.RR.R  3.00 monthly            KPRA     RR            R
#> 34: 2026-01-01 M.KPRA.RR.R  3.00 monthly            KPRA     RR            R
#> 35: 2026-02-01 M.KPRA.RR.R  3.00 monthly            KPRA     RR            R
#> 36: 1975-07-15 A.KPRA.RR.R -0.92  annual            KPRA     RR            R
#> 37: 1975-07-16 A.KPRA.RR.R  0.33  annual            KPRA     RR            R
#> 38: 1975-07-17 A.KPRA.RR.R  2.54  annual            KPRA     RR            R
#> 39: 1975-07-18 A.KPRA.RR.R  3.50  annual            KPRA     RR            R
#> 40: 1975-07-19 A.KPRA.RR.R  3.29  annual            KPRA     RR            R
#> 41: 2025-10-01 M.KPRA.SD.R  4.00 monthly            KPRA     SD            R
#> 42: 2025-11-01 M.KPRA.SD.R  4.00 monthly            KPRA     SD            R
#> 43: 2025-12-01 M.KPRA.SD.R  4.00 monthly            KPRA     SD            R
#> 44: 2026-01-01 M.KPRA.SD.R  4.00 monthly            KPRA     SD            R
#> 45: 2026-02-01 M.KPRA.SD.R  4.00 monthly            KPRA     SD            R
#>           date         key value    freq instrument_type  tenor unit_measure
#>         <Date>      <char> <num>  <char>          <char> <char>       <char>
# }