Fetch Sveriges Riksbank (SRb) data

Retrieve time series data from the Sveriges Riksbank SWEA API.

Usage

srb_data(series, start_date = NULL, end_date = NULL)

Source

https://developer.api.riksbank.se/

Arguments

series

(character(1))
The series ID to query. See srb_series() for available series.

start_date

(NULL | Date(1) | character(1))
Start date of the data (e.g., "2024-01-01"). If NULL, all available data is returned. Default NULL.

end_date

(NULL | Date(1) | character(1))
End date of the data, in the same format as start_date. If NULL, data up to the latest available date is returned. Default NULL.

Value

A data.table::data.table() with the requested data.

See also

Examples

# \donttest{
# fetch USD/SEK exchange rate
srb_data("SEKUSDPMI", start_date = "2024-01-01")
#>            date    series    value
#>          <Date>    <char>    <num>
#>   1: 2024-01-02 SEKUSDPMI 10.18118
#>   2: 2024-01-03 SEKUSDPMI 10.24956
#>   3: 2024-01-04 SEKUSDPMI 10.21684
#>   4: 2024-01-05 SEKUSDPMI 10.28752
#>   5: 2024-01-08 SEKUSDPMI 10.24073
#>  ---                              
#> 547: 2026-03-10 SEKUSDPMI  9.11090
#> 548: 2026-03-11 SEKUSDPMI  9.19981
#> 549: 2026-03-12 SEKUSDPMI  9.27583
#> 550: 2026-03-13 SEKUSDPMI  9.37130
#> 551: 2026-03-16 SEKUSDPMI  9.38230

# fetch EUR/SEK exchange rate
srb_data("SEKEURPMI", start_date = "2024-01-01")
#>            date    series   value
#>          <Date>    <char>   <num>
#>   1: 2024-01-02 SEKEURPMI 11.1545
#>   2: 2024-01-03 SEKEURPMI 11.1915
#>   3: 2024-01-04 SEKEURPMI 11.1905
#>   4: 2024-01-05 SEKEURPMI 11.2350
#>   5: 2024-01-08 SEKEURPMI 11.2095
#>  ---                             
#> 547: 2026-03-10 SEKEURPMI 10.6060
#> 548: 2026-03-11 SEKEURPMI 10.6543
#> 549: 2026-03-12 SEKEURPMI 10.7108
#> 550: 2026-03-13 SEKEURPMI 10.7545
#> 551: 2026-03-16 SEKEURPMI 10.7690
# }